Observer for Treatment reference variable

Updates the reference and comparison Treatments when the selected Treatment variable changes

arm_ref_comp_observer(
  session,
  input,
  id_ref,
  id_comp,
  id_arm_var,
  datasets,
  dataname = "ADSL",
  arm_ref_comp,
  module,
  on_off = reactive(TRUE)
)

Arguments

session

(environment) shiny session

input

(character) shiny input

id_ref

(character) id of reference Treatment input ui element

id_comp

(character) id of comparison group input ui element

id_arm_var

(character) id of Treatment variable input ui element

datasets

(FilteredData) object from the module

dataname

(character) dataset name

arm_ref_comp

(unknown) Treatment reference and compare variables provided as a nested list where each Treatment variable corresponds a list specifying the default levels for the reference and comparison treatments.

module

(character) name of the module where this is called (this is only used to produce more informative error messages)

on_off

(logical) A reactive that can be used to stop the whole observer if FALSE.

Examples

ds <- teal:::get_dummy_datasets()
#> Loading required package: scda
#> 

arm_ref_comp <- list(ARMCD = list(ref = "ARM A", comp = c("ARM B")))
arm_var <- teal::choices_selected(c("ARM", "ARMCD"), "ARM")
if (FALSE) {
shinyApp(
  ui = fluidPage(
    teal.widgets::optionalSelectInput(
      "arm",
      "Treatment Variable",
      choices = arm_var$choices,
      selected = arm_var$selected
    ),
    selectInput(
      "ref",
      "Reference Treatment",
      choices = NULL,
      multiple = TRUE
    ),
    selectInput(
      "comp",
      "Comparison Group",
      choices = NULL,
      multiple = TRUE
    )
  ),
  server = function(input, output, session) {
    isolate({
      teal.modules.clinical:::arm_ref_comp_observer(
        session,
        input,
        id_ref = "ref",
        id_comp = "comp",
        id_arm_var = "arm",
        datasets = ds,
        arm_ref_comp = arm_ref_comp,
        module = "example"
      )
    })
  }
)
}