teal Modules for Standard Clinical Outputs

Provides teal modules for the standard clinical trials outputs. The teal modules add an encoding panel to interactively change the encodings within teal.

See also

Author

Maintainer: Dawid Kaledkowski dawid.kaledkowski@roche.com

Authors:

Joe Zhu joe.zhu@roche.com (ORCID)

Jana Stoilova jana.stoilova@roche.com

Davide Garolini davide.garolini@roche.com

Emily de la Rua emily.de_la_rua@contractors.roche.com

Abinaya Yogasekaram abinaya.yogasekaram@contractors.roche.com

Mahmoud Hallal mahmoud.hallal@roche.com

Rosemary Li li.yaqiong@gene.com

Heng Wang wang.heng@gene.com

Pawel Rucki pawel.rucki@roche.com

Nikolas Burkoff

Konrad Pagacz

Other contributors:

Vaakesan Sundrelingam [contributor]

Francois Collin [contributor]

Imanol Zubizarreta [contributor]

F. Hoffmann-La Roche AG [copyright holder, funder]