Template: Binary Outcome

Creates a valid expression for binary outcome analysis.

template_binary_outcome(
  dataname,
  parentname,
  arm_var,
  paramcd,
  ref_arm = NULL,
  comp_arm = NULL,
  compare_arm = FALSE,
  combine_comp_arms = FALSE,
  aval_var = "AVALC",
  show_rsp_cat = TRUE,
  responder_val = c("Complete Response (CR)", "Partial Response (PR)"),
  control = list(global = list(method = "waldcc", conf_level = 0.95), unstrat =
    list(method_ci = "waldcc", method_test = "schouten", odds = TRUE), strat =
    list(method_ci = "cmh", method_test = "cmh", strat = NULL)),
  add_total = FALSE,
  basic_table_args = teal.devel::basic_table_args()
)

Arguments

dataname

(character)
analysis data used in teal module.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var

(character)
variable names that can be used as arm_var.

paramcd

(character)
response parameter value to use in the table title.

ref_arm

(character)
the level of reference arm in case of arm comparison.

comp_arm

(character)
the level of comparison arm in case of arm comparison.

compare_arm

(logical)
triggers the comparison between study arms.

combine_comp_arms

(logical)
triggers the combination of comparison arms.

aval_var

(character)
name of the analysis variable.

show_rsp_cat

(logical)
display the multinomial response estimations.

responder_val

(character)
the short label for observations to translate AVALC into responder / non-responder.

control

(list)
list of settings for the analysis.

add_total

(logical)
whether to include column with total number of patients.

basic_table_args

optional, (basic_table_args)
object created by teal.devel::basic_table_args() with settings for the module table. The argument is merged with option teal.basic_table_args and with default module arguments (hard coded in the module body).
For more details, see the help vignette:
vignette("Custom basic_table arguments module", package = "teal.devel").

See also

Examples

if (FALSE) {
# Preparation of the test case.
library(dplyr)
library(scda)
library(tern)
adsl <- synthetic_cdisc_data("latest")$adsl
adrs <- synthetic_cdisc_data("latest")$adrs

# Generate an expression for the analysis of responders.
a <- template_binary_outcome(
  dataname = "adrs",
  parentname = "adsl",
  arm_var = "ARMCD",
  paramcd = "BESRSPI",
  ref_arm = "ARM A",
  comp_arm = c("ARM B"),
  compare_arm = TRUE,
  show_rsp_cat = TRUE
)

b <- mapply(expr = a, FUN = eval)
b$data
b$layout
b$table
}