Template: Event rates adjusted for patient-years

Template: Event rates adjusted for patient-years

template_events_patyear(
  dataname,
  parentname,
  arm_var,
  events_var,
  aval_var = "AVAL",
  add_total = TRUE,
  control = control_incidence_rate(),
  drop_arm_levels = TRUE
)

Arguments

dataname
(character)
analysis data used in teal module.

parentname
(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var
(character)
variable names that can be used as arm_var.

events_var
(integer)
number of observed events.

aval_var
(character)
name of the analysis variable.

add_total
(logical)
whether to include column with total number of patients.

control
(list)
list of settings for the analysis.

drop_arm_levels
(logical)
drop the unused arm_var levels. When TRUE, arm_var levels are set to those used in the dataname dataset. When FALSE, arm_var levels are set to those used in the parantname dataset.

See also