Template: Adverse Events Summary

Template: Adverse Events Summary

Usage

template_events_summary(
  anl_name,
  parentname,
  arm_var,
  dthfl_var = "DTHFL",
  dcsreas_var = "DCSREAS",
  flag_var_anl = NULL,
  flag_var_aesi = NULL,
  aeseq_var = "AESEQ",
  llt = "AEDECOD",
  add_total = TRUE,
  count_subj = TRUE,
  count_pt = TRUE,
  count_events = TRUE
)

Arguments

anl_name

(character)
analysis data used in teal module.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var

(character)
variable names that can be used as arm_var.

dthfl_var

(character)
variable for subject death flag from parentname. Records with `"Y"`` are summarized in the table row for "Total number of deaths".

dcsreas_var

(character)
variable for study discontinuation reason from parentname. Records with "ADVERSE EVENTS" are summarized in the table row for "Total number of patients withdrawn from study due to an AE".

flag_var_anl

(character)
flag variable from dataset used to count adverse event sub-groups (e.g. Serious events, Related events, etc.). Variable labels are used as table row names if they exist.

flag_var_aesi

(character)
flag variable from dataset used to count adverse event special interest groups. All flag variables must be of type logical. Variable labels are used as table row names if they exist.

aeseq_var

(character)
variable for adverse events sequence number from dataset. Used for counting total number of events.

llt

(character)
name of the variable with low level term for events.

add_total

(logical)
whether to include column with total number of patients.

count_subj

(logical)
w whether to show count of unique subjects based on USUBJID. Only applies if event flag variables are provided.

count_pt

(logical)
whether to show count of unique preferred terms based on llt. Only applies if event flag variables are provided.

count_events

(logical)
whether to show count of events based on aeseq_var. Only applies if event flag variables are provided.

See also