Template: Mixed Model Repeated Measurements (MMRM) analysis

Template: Mixed Model Repeated Measurements (MMRM) analysis

template_fit_mmrm(
  parentname,
  dataname,
  aval_var,
  arm_var,
  ref_arm,
  comp_arm = NULL,
  combine_comp_arms = FALSE,
  id_var,
  visit_var,
  cov_var,
  conf_level = 0.95,
  cor_struct = "unstructured",
  weights_emmeans = "proportional",
  optimizer = "automatic",
  parallel = FALSE
)

template_mmrm_tables(
  parentname,
  dataname,
  fit_name,
  arm_var,
  ref_arm,
  visit_var,
  paramcd,
  show_relative = c("increase", "reduction", "none"),
  table_type = "t_mmrm_cov",
  basic_table_args = teal.widgets::basic_table_args()
)

template_mmrm_plots(
  fit_name,
  lsmeans_plot = list(select = c("estimates", "contrasts"), width = 0.6, show_pval =
    FALSE),
  diagnostic_plot = list(type = "fit-residual", z_threshold = NULL),
  ggplot2_args = teal.widgets::ggplot2_args()
)

Arguments

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

dataname

(character)
analysis data used in teal module.

aval_var

(character)
name of the analysis variable.

arm_var

(character)
variable names that can be used as arm_var.

ref_arm

(character)
the level of reference arm in case of arm comparison.

comp_arm

(character)
the level of comparison arm in case of arm comparison.

combine_comp_arms

(logical)
triggers the combination of comparison arms.

id_var

(character)
the variable name for subject id.

visit_var

(character)
variable names that can be used as visit variable. Must be a factor in dataname.

cov_var

(character)
names of the covariates variables.

conf_level

(numeric)
value for the confidence level within the range of (0, 1).

cor_struct

a string specifying the correlation structure, defaults to "unstructured". See the details.

weights_emmeans

argument from emmeans::emmeans(), "proportional" by default.

optimizer

a string specifying the optimization algorithm which should be used. By default, "automatic" will (if necessary) try all possible optimization algorithms and choose the best result. If another algorithm is chosen and does not give a valid result, an error will occur.

parallel

flag that controls whether "automatic" optimizer search can use available free cores on the machine (not default).

fit_name

name of fitted MMRM object

paramcd

(character)
variable value designating the studied parameter.

show_relative

should the "reduction" (control - treatment, default) or the "increase" (treatment - control) be shown for the relative change from baseline

table_type

(character)
type of table to output.

basic_table_args

optional, (basic_table_args)
object created by teal.widgets::basic_table_args() with settings for the module table. The argument is merged with option teal.basic_table_args and with default module arguments (hard coded in the module body).

For more details, see the vignette: vignette("custom-basic-table-arguments", package = "teal.widgets").

lsmeans_plot

a list of controls for LS means plot. See more tern.mmrm::g_mmrm_lsmeans()

diagnostic_plot

a list of controls for diagnostic_plot. See more tern.mmrm::g_mmrm_diagnostic()

ggplot2_args

optional, (ggplot2_args)
object created by teal.widgets::ggplot2_args() with settings for the module plot. The argument is merged with option teal.ggplot2_args and with default module arguments (hard coded in the module body).

For more details, see the vignette: vignette("custom-ggplot2-arguments", package = "teal.widgets").

Functions

template_mmrm_tables:

template_mmrm_plots:

See also