Template: Line Plot

Template: Line Plot

template_g_lineplot(
  dataname = "ANL",
  strata = "ARM",
  x = "AVISIT",
  y = "AVAL",
  y_unit = "AVALU",
  paramcd = "PARAMCD",
  param = "ALT",
  mid = "mean",
  interval = "mean_ci",
  whiskers = c("mean_ci_lwr", "mean_ci_upr"),
  table = c("n", "mean_sd", "median", "range"),
  mid_type = "pl",
  conf_level = 0.95,
  incl_screen = TRUE,
  mid_point_size = 2,
  table_font_size = 4,
  title = "Line Plot",
  y_lab = "",
  ggplot2_args = teal.devel::ggplot2_args()
)

Arguments

dataname

(character)
analysis data used in teal module.

strata

(character scalar or NA) strata variable name.

x

(character scalar) x variable name.

y

(character scalar) y variable name.

y_unit

(character scalar or NA) y_unit variable name.

paramcd

(character scalar or NA) paramcd variable name.

param

(character)
parameter chosen to filter the data by.

mid

(character or NULL)
names of the statistics that will be plotted as midpoints. All the statistics indicated in mid variable must be present in the object returned by sfun, and be of a double or numeric type vector of length one.

interval

(character or NULL)
names of the statistics that will be plotted as intervals. All the statistics indicated in interval variable must be present in the object returned by sfun, and be of a double or numeric type vector of length two.

whiskers

(character)
names of the interval whiskers that will be plotted. Must match the names attribute of the interval element in the list returned by sfun. It is possible to specify one whisker only, lower or upper.

table

(character or NULL)
names of the statistics that will be displayed in the table below the plot. All the statistics indicated in table variable must be present in the object returned by sfun.

mid_type

(character scalar)
controls the type of the mid plot, it can be point (p), line (l), or point and line (pl).

conf_level

(numeric)
value for the confidence level within the range of (0, 1).

incl_screen

(logical)
should the screening visit be included.

mid_point_size

(integer or double)
controls the font size of the point for mid plot.

table_font_size

(integer or double)
controls the font size of values in the table.

title

(character scalar)
plot title.

y_lab

(character scalar)
y-axis label. If it equals to NULL, then no label will be added.

ggplot2_args

optional, (ggplot2_args)
object created by teal.devel::ggplot2_args() with settings for the module plot. For this module, this argument will only accept ggplot2_args object with labs list of following child elements: title, subtitle, caption, y, lty. No other elements would be taken into account. The argument is merged with option teal.ggplot2_args and with default module arguments (hard coded in the module body).

For more details, see the vignette: vignette("custom-ggplot2-arguments", package = "teal.devel").

See also