Template: Logistic Regression

Creates an expression for logistic regressions.

template_logistic(
  dataname,
  arm_var,
  aval_var,
  cov_var,
  interaction_var,
  ref_arm,
  comp_arm,
  topleft = "Logistic Regression",
  conf_level = 0.95,
  combine_comp_arms = FALSE,
  responder_val = c("CR", "PR"),
  at = NULL
)

Arguments

dataname
(character)
analysis data used in teal module.

arm_var
(character)
variable names that can be used as arm_var.

aval_var
(character)
name of the analysis variable.

cov_var
(character)
names of the covariates variables.

interaction_var
(character)
names of the variables that can be used for interaction variable selection.

ref_arm
(character)
the level of reference arm in case of arm comparison.

comp_arm
(character)
the level of comparison arm in case of arm comparison.

topleft
(character)
the top-left annotation in the table.

conf_level
(numeric)
value for the confidence level within the range of (0, 1).

combine_comp_arms
(logical)
triggers the combination of comparison arms.

responder_val
(character)
values of the responder variable corresponding with a successful response.

at
optional, (NULL or numeric)
values for the interaction variable. Otherwise the median is used.

See also