Template: Shift by Arm

Template: Shift by Arm

Usage

template_shift_by_arm(
  dataname,
  parentname,
  arm_var = "ARM",
  paramcd = "PARAMCD",
  visit_var = "AVISIT",
  treatment_flag_var = "ONTRTFL",
  treatment_flag = "Y",
  aval_var = "ANRIND",
  base_var = "BNRIND",
  na.rm = FALSE,
  na_level = "<Missing>",
  add_total = FALSE,
  total_label = "All Patients",
  basic_table_args = teal.widgets::basic_table_args()
)

Arguments

dataname

(character)
analysis data used in teal module.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var

(character)
variable names that can be used as arm_var.

paramcd

(character)
variable value designating the studied parameter.

visit_var

(character)
variable names that can be used as visit variable. Must be a factor in dataname.

treatment_flag_var

(character)
name of the on treatment flag variable.

treatment_flag

(character)
name of the value indicating on treatment records in treatment_flag_var.

aval_var

(character)
the variable name for the analysis reference range indicator.

base_var

(character)
the variable name for the baseline reference range indicator.

na.rm

(logical)
whether NA values should be removed prior to analysis.

na_level

(string)
used to replace all NA or empty values in character or factor variables in the data.

add_total

(logical)
whether to include row with total number of patients.

basic_table_args

optional, (basic_table_args)
object created by teal.widgets::basic_table_args() with settings for the module table. The argument is merged with option teal.basic_table_args and with default module arguments (hard coded in the module body).

For more details, see the vignette: vignette("custom-basic-table-arguments", package = "teal.widgets").

See also