Adverse Events Table by Standardized MedDRA Query

Adverse Events Table by Standardized MedDRA Query

template_smq(
  dataname,
  parentname,
  arm_var,
  llt = "AEDECOD",
  add_total = TRUE,
  sort_criteria = c("freq_desc", "alpha"),
  drop_arm_levels = TRUE,
  na_level = "<Missing>",
  smq_varlabel = "Standardized MedDRA Query",
  baskets = c("SMQ01NAM", "SMQ02NAM", "CQ01NAM"),
  id_var = "USUBJID",
  basic_table_args = teal.widgets::basic_table_args()
)

Arguments

dataname

(character)
analysis data used in teal module.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var

(character)
variable names that can be used as arm_var.

llt

(character)
name of the variable with low level term for events.

add_total

(logical)
whether to include column with total number of patients.

sort_criteria

(character)
how to sort the final table. Default option freq_desc sorts by decreasing total number of patients with event. Alternative option alpha sorts events alphabetically.

drop_arm_levels

(logical)
drop the unused arm_var levels. When TRUE, arm_var levels are set to those used in the dataname dataset. When FALSE, arm_var levels are set to those used in the parantname dataset.

na_level

(string)
used to replace all NA or empty values in character or factor variables in the data.

smq_varlabel

(character)
label of the new column SMQ created by tern::h_stack_by_baskets().

baskets

(character)
variable names of the selected Standardized/Customized queries

id_var

(character)
the variable name for subject id.

basic_table_args

optional, (basic_table_args)
object created by teal.widgets::basic_table_args() with settings for the module table. The argument is merged with option teal.basic_table_args and with default module arguments (hard coded in the module body).

For more details, see the vignette: vignette("custom-basic-table-arguments", package = "teal.widgets").

See also