Template: Summary of Variables

Template: Summary of Variables

template_summary(
  dataname,
  parentname,
  arm_var,
  sum_vars,
  show_labels = c("default", "visible", "hidden"),
  add_total = TRUE,
  var_labels = character(),
  na.rm = FALSE,
  na_level = "<Missing>",
  denominator = c("N", "n", "omit"),
  drop_arm_levels = TRUE
)

Arguments

dataname
(character)
analysis data used in teal module.

parentname
(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var
(character)
variable names that can be used as arm_var.

sum_vars
(character)
names of the variables that should be summarized.

show_labels
(character)
defines whether the labels for sum_vars should display. For details see rtables::analyze().

add_total
(logical)
whether to include column with total number of patients.

var_labels
optional, (named character)
variable labels for relabeling the analysis variables.

na.rm
(logical)
whether NA values should be removed prior to analysis.

na_level
(string)
used to replace all NA or empty values in character or factor variables in the data.

denominator
(character)
chooses how percentages are calculated. With option N, the reference population from the column total is used as the denominator. With option n, the number of non-missing records in this row and column intersection is used as the denominator. If omit is chosen, then the percentage is omitted.

drop_arm_levels
(logical)
drop the unused arm_var levels. When TRUE, arm_var levels are set to those used in the dataname dataset. When FALSE, arm_var levels are set to those used in the parantname dataset.

See also