Template: Time-to-Event

Creates a valid expression for time-to-event analysis.

template_tte(
  dataname = "ANL",
  parentname = "ADSL_FILTERED",
  arm_var = "ARM",
  paramcd,
  ref_arm = NULL,
  comp_arm = NULL,
  compare_arm = FALSE,
  combine_comp_arms = FALSE,
  aval_var = "AVAL",
  cnsr_var = "CNSR",
  strata_var = NULL,
  time_points = NULL,
  time_unit_var = "AVALU",
  event_desc_var = "EVNTDESC",
  control = control_tte(),
  add_total = FALSE,
  basic_table_args = teal.widgets::basic_table_args()
)

Arguments

dataname

(character)
analysis data used in teal module.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var

(character)
variable names that can be used as arm_var.

paramcd

(character)
endpoint parameter value to use in the table title.

ref_arm

(character)
the level of reference arm in case of arm comparison.

comp_arm

(character)
the level of comparison arm in case of arm comparison.

compare_arm

(logical)
triggers the comparison between study arms.

combine_comp_arms

(logical)
triggers the combination of comparison arms.

aval_var

(character)
name of the analysis variable.

cnsr_var

(character)
name of the censoring variable.

strata_var

(character)
names of the variables for stratified analysis.

time_points

(character)
time points that can be used in tern::surv_timepoint().

time_unit_var

(character)
name of the variable representing time units.

event_desc_var

(character)
name of the variable with events description.

control

(list)
list of settings for the analysis, see control_tte().

add_total

(logical)
whether to include column with total number of patients.

basic_table_args

optional, (basic_table_args)
object created by teal.widgets::basic_table_args() with settings for the module table. The argument is merged with option teal.basic_table_args and with default module arguments (hard coded in the module body).

For more details, see the vignette: vignette("custom-basic-table-arguments", package = "teal.widgets").

See also