Teal Module: Forest Survival Plot teal Module

This teal module produces a grid style Forest plot for time-to-event data with ADaM structure

tm_g_forest_tte(
  label,
  dataname,
  parentname = ifelse(is(arm_var, "data_extract_spec"), datanames_input(arm_var),
    "ADSL"),
  arm_var,
  arm_ref_comp = NULL,
  subgroup_var,
  paramcd,
  strata_var,
  aval_var = choices_selected(variable_choices(dataname, "AVAL"), "AVAL", fixed = TRUE),
  cnsr_var = choices_selected(variable_choices(dataname, "CNSR"), "CNSR", fixed = TRUE),
  conf_level = choices_selected(c(0.95, 0.9, 0.8), 0.95, keep_order = TRUE),
  time_unit_var = choices_selected(variable_choices(dataname, "AVALU"), "AVALU", fixed
    = TRUE),
  fixed_symbol_size = TRUE,
  plot_height = c(700L, 200L, 2000L),
  plot_width = c(980L, 500L, 2000L),
  pre_output = NULL,
  post_output = NULL
)

Arguments

label
(character)
menu item label of the module in the teal app

dataname
(character)
analysis data used in teal module.

parentname
(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable names that can be used as arm_var. It defines the grouping variable(s) in the results table. If there are two elements selected for arm_var, second variable will be nested under the first variable.

arm_ref_comp
optional, (list)
If specified it must be a named list with each element corresponding to an arm variable in ADSL and the element must be another list (possibly with delayed teal::variable_choices() or delayed teal::value_choices() with the elements named ref and comp that the defined the default reference and comparison arms when the arm variable is changed.

subgroup_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable names that can be used as the default subgroups.

paramcd
(teal::choices_selected() or teal::data_extract_spec())
variable value designating the studied parameter.

strata_var
(teal::choices_selected() or teal::data_extract_spec())
names of the variables for stratified analysis.

aval_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for the analysis variable.

cnsr_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for the censoring variable.

conf_level
(teal::choices_selected())
object with all available choices and preselected option for the confidence level, each within range of (0, 1).

time_unit_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for the time unit variable.

fixed_symbol_size
(logical)
When (TRUE), the same symbol size is used for plotting each estimate. Otherwise, the symbol size will be proportional to the sample size in each each subgroup.

plot_height
optional, (numeric)
a vector of length three with c(value, min, max). Specifies the height of the main plot and renders a slider on the plot to interactively adjust the plot height.

plot_width
optional, (numeric)
a vector of length three with c(value, min, max). Specifies the width of the main plot and renders a slider on the plot to interactively adjust the plot width.

pre_output
optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output
optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

Examples

library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl
ADTTE <- synthetic_cdisc_data("latest")$adtte

ADSL$RACE <- droplevels(ADSL$RACE)

arm_ref_comp = list(
  ARM = list(
    ref = "B: Placebo",
    comp = c("A: Drug X", "C: Combination")
  ),
  ARMCD = list(
    ref = "ARM B",
    comp = c("ARM A", "ARM C")
  )
)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL,
      code = 'ADSL <- synthetic_cdisc_data("latest")$adsl
              ADSL$RACE <- droplevels(ADSL$RACE)'),
    cdisc_dataset("ADTTE", ADTTE, code = 'ADTTE <- synthetic_cdisc_data("latest")$adtte'),
    check = TRUE
  ),
  modules = root_modules(
    tm_g_forest_tte(
       label = "Forest Survival",
       dataname = "ADTTE",
       arm_var = choices_selected(
         variable_choices(ADSL, c("ARM", "ARMCD")),
         "ARMCD"
       ),
       arm_ref_comp = arm_ref_comp,
       paramcd = choices_selected(
         value_choices(ADTTE, "PARAMCD", "PARAM"),
         "OS"
       ),
       subgroup_var = choices_selected(
         variable_choices(ADSL, names(ADSL)),
         c("BMRKR2", "SEX")
       ),
       strata_var = choices_selected(
         variable_choices(ADSL, c("STRATA1", "STRATA2")),
         "STRATA2"
       ),
       plot_height = c(600, 200, 2000)
    )
  )
)
#> 2021-12-08 09:16:26.0079 pid:866 teal.modules.clinical fun:tm_g_forest_tte [INFO] Initializing tm_g_forest_tte

if (FALSE) {
shinyApp(app$ui, app$server)
}