Teal Module: Abnormality Summary Table

Teal Module: Abnormality Summary Table

tm_t_abnormality(
  label,
  dataname,
  parentname = ifelse(is(arm_var, "data_extract_spec"), datanames_input(arm_var),
    "ADSL"),
  arm_var,
  by_vars,
  grade,
  abnormal = list(low = c("LOW", "LOW LOW"), high = c("HIGH", "HIGH HIGH")),
  id_var = choices_selected(variable_choices(dataname, subset = "USUBJID"), selected =
    "USUBJID", fixed = TRUE),
  baseline_var = choices_selected(variable_choices(dataname, subset = "BNRIND"),
    selected = "BNRIND", fixed = TRUE),
  treatment_flag_var = choices_selected(variable_choices(dataname, subset = "ONTRTFL"),
    selected = "ONTRTFL", fixed = TRUE),
  treatment_flag = choices_selected(value_choices(dataname, "ONTRTFL"), selected = "Y",
    fixed = TRUE),
  add_total = TRUE,
  exclude_base_abn = FALSE,
  drop_arm_levels = TRUE,
  pre_output = NULL,
  post_output = NULL,
  na_level = "<Missing>"
)

Arguments

label
(character)
menu item label of the module in the teal app

dataname
(character)
analysis data used in teal module.

parentname
(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable names that can be used as arm_var. It defines the grouping variable(s) in the results table. If there are two elements selected for arm_var, second variable will be nested under the first variable.

by_vars
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable names used to split the summary by rows.

grade
(teal::choices_selected() or teal::data_extract_spec)
object with all available choices and preselected option for variable names that can be used to specify the abnormality grade. Variable must be factor.

abnormal
(named list)
defined by user to indicate what abnormalities are to be displayed.

id_var
(teal::choices_selected() or teal::data_extract_spec())
object specifying the variable name for subject id.

baseline_var
(teal::choices_selected() or teal::data_extract_spec)
variable for baseline abnormality grade.

treatment_flag_var
(teal::choices_selected() or teal::data_extract_spec)
on treatment flag variable.

treatment_flag
(teal::choices_selected() or teal::data_extract_spec)
value indicating on treatment records in treatment_flag_var.

add_total
(logical)
whether to include column with total number of patients.

exclude_base_abn
(logical)
whether to exclude patients who had abnormal values at baseline.

drop_arm_levels
(logical)
drop the unused arm_var levels. When TRUE, arm_var levels are set to those used in the dataname dataset. When FALSE, arm_var levels are set to those used in the parantname dataset.

pre_output
optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output
optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

na_level
(character)
the NA level in the input dataset, default to "<Missing>".

Note

Patients with the same abnormality at baseline as on the treatment visit can be excluded in accordance with GDSR specifications by using exclude_base_abn.

Examples

library(scda)
library(dplyr)

adsl <- synthetic_cdisc_data("latest")$adsl
adlb <- synthetic_cdisc_data("latest")$adlb %>%
  mutate(
    ONTRTFL = case_when(
      AVISIT %in% c("SCREENING", "BASELINE") ~ "",
      TRUE ~ "Y"
    )
  ) %>%
  var_relabel(
    ONTRTFL = "On Treatment Record Flag"
  )

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", adsl, code = "ADSL <- synthetic_cdisc_data('latest')$adsl"),
    cdisc_dataset("ADLB", adlb,
      code = "ADLB <- synthetic_cdisc_data('latest')$adlb %>%
                mutate(
                  ONTRTFL = case_when(
                    AVISIT %in% c('SCREENING', 'BASELINE') ~ '',
                    TRUE ~ 'Y'
                  )
                ) %>%
                var_relabel(
                  ONTRTFL = 'On Treatment Record Flag'
                )"
    ),
    check = TRUE
  ),
  modules = root_modules(
    tm_t_abnormality(
      label = "Abnormality Table",
      dataname = "ADLB",
      arm_var = choices_selected(
        choices = variable_choices(adsl, subset = c("ARM", "ARMCD")),
        selected = "ARM"
      ),
      add_total = FALSE,
      by_vars = choices_selected(
        choices = variable_choices(adlb, subset = c("LBCAT", "PARAM", "AVISIT")),
        selected = c("LBCAT", "PARAM"),
        keep_order = TRUE
      ),
      grade = choices_selected(
        choices = variable_choices(adlb, subset = "ANRIND"),
        selected = "ANRIND",
        fixed = TRUE
      ),
      abnormal = list(low = "LOW", high = "HIGH"),
      exclude_base_abn = FALSE
    )
  )
)

if (FALSE) {
shinyApp(app$ui, app$server)
}