Teal Module: Laboratory test results with highest grade post-baseline

Teal Module: Laboratory test results with highest grade post-baseline

tm_t_abnormality_by_worst_grade(
  label,
  dataname,
  parentname = ifelse(inherits(arm_var, "data_extract_spec"),
    teal.devel::datanames_input(arm_var), "ADSL"),
  arm_var,
  id_var = choices_selected(variable_choices(dataname, subset = "USUBJID"), selected =
    "USUBJID", fixed = TRUE),
  paramcd,
  atoxgr_var = choices_selected(variable_choices(dataname, subset = "ATOXGR"), selected
    = "ATOXGR", fixed = TRUE),
  worst_high_flag_var = choices_selected(variable_choices(dataname, subset =
    "WGRHIFL"), selected = "WGRHIFL", fixed = TRUE),
  worst_low_flag_var = choices_selected(variable_choices(dataname, subset = "WGRLOFL"),
    selected = "WGRLOFL", fixed = TRUE),
  worst_flag_indicator = choices_selected(value_choices(dataname, var_choices =
    "WGRLOFL"), selected = "Y", fixed = TRUE),
  add_total = TRUE,
  drop_arm_levels = TRUE,
  pre_output = NULL,
  post_output = NULL,
  basic_table_args = teal.devel::basic_table_args()
)

Arguments

label

(character)
menu item label of the module in the teal app.

dataname

(character)
analysis data used in teal module.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var

(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable names that can be used as arm_var. It defines the grouping variable(s) in the results table. If there are two elements selected for arm_var, second variable will be nested under the first variable.

id_var

(teal::choices_selected() or teal::data_extract_spec())
object specifying the variable name for subject id.

paramcd

(teal::choices_selected() or teal::data_extract_spec())
variable value designating the studied parameter.

atoxgr_var

(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable names that can be used as Analysis Toxicity Grade.

worst_high_flag_var

(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable names that can be used as Worst High Grade flag.

worst_low_flag_var

(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable names that can be used as Worst Low Grade flag.

worst_flag_indicator

(teal::choices_selected() or teal::data_extract_spec())
value indicating worst grade.

add_total

(logical)
whether to include column with total number of patients.

drop_arm_levels

(logical)
drop the unused arm_var levels. When TRUE, arm_var levels are set to those used in the dataname dataset. When FALSE, arm_var levels are set to those used in the parantname dataset.

pre_output

optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output

optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

basic_table_args

optional, (basic_table_args)
object created by teal.devel::basic_table_args() with settings for the module table. The argument is merged with option teal.basic_table_args and with default module arguments (hard coded in the module body).

For more details, see the vignette: vignette("custom-basic-table-arguments", package = "teal.devel").

See also

Examples

library(scda)
library(dplyr)

adsl <- synthetic_cdisc_data("latest")$adsl
adlb <- synthetic_cdisc_data("latest")$adlb %>%
  filter(!AVISIT %in% c("SCREENING", "BASELINE"))

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", adsl, code = 'ADSL <- synthetic_cdisc_data("latest")$adsl'),
    cdisc_dataset("ADLB", adlb,
      code = "ADLB <- synthetic_cdisc_data('latest')$adlb %>%
        filter(!AVISIT %in% c('SCREENING', 'BASELINE'))"
    ),
    check = TRUE
  ),
  modules = root_modules(
    tm_t_abnormality_by_worst_grade(
      label = "Laboratory test results with highest grade post-baseline",
      dataname = "ADLB",
      arm_var = choices_selected(
        choices = variable_choices(adsl, subset = c("ARM", "ARMCD")),
        selected = "ARM"
      ),
      paramcd = choices_selected(
        choices = value_choices(adlb, "PARAMCD", "PARAM"),
        selected = c("ALT", "CRP", "IGA")
      ),
      add_total = FALSE
    )
  ),
  filter = (
    list(
      ADSL = list(SAFFL = "Y"),
      ADLB = list(ONTRTFL = "Y")
    )
  )
)
#> [INFO] 2022-02-07 10:30:53.1308 pid:2152 token:[] teal.modules.clinical Initializing tm_t_abnormality_by_worst_grade
if (FALSE) {
shinyApp(app$ui, app$server)
}