Teal module: Exposure Table for Risk management plan

Teal module: Exposure Table for Risk management plan

tm_t_exposure(
  label,
  dataname,
  parentname = ifelse(is(col_by_var, "data_extract_spec"), datanames_input(col_by_var),
    "ADSL"),
  row_by_var,
  col_by_var,
  paramcd = choices_selected(choices = value_choices(dataname, "PARAMCD", "PARAM"),
    selected = "TDURD"),
  id_var = choices_selected(variable_choices(dataname, subset = "USUBJID"), selected =
    "USUBJID", fixed = TRUE),
  parcat,
  aval_var = choices_selected(variable_choices(dataname, subset = "AVAL"), selected =
    "AVAL", fixed = TRUE),
  avalu_var = choices_selected(variable_choices(dataname, subset = "AVALU"), selected =
    "AVALU", fixed = TRUE),
  add_total,
  na_level = "<Missing>",
  pre_output = NULL,
  post_output = NULL
)

Arguments

label
(character)
menu item label of the module in the teal app

dataname
(character)
analysis data used in teal module.

parentname
(character)
parent analysis data used in teal module, usually this refers to ADSL.

row_by_var
(teal::choices_selected() or teal::data_extract_spec)
object with all available choices and preselected option for variable names that can be used to split rows.

col_by_var
(teal::choices_selected() or teal::data_extract_spec)
object with all available choices and preselected option for variable names that can be used to split columns.

paramcd
(teal::choices_selected() or teal::data_extract_spec())
variable value designating the studied parameter.

id_var
(teal::choices_selected() or teal::data_extract_spec())
object specifying the variable name for subject id.

parcat
(teal::choices_selected() or teal::data_extract_spec)
object with all available choices and preselected option for parameter category values.

aval_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for the analysis variable.

avalu_var
(teal::choices_selected() or teal::data_extract_spec)
object with the analysis unit variable.

add_total
(logical)
whether to include column with total number of patients.

na_level
(string)
used to replace all NA or empty values in character or factor variables in the data.

pre_output
optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output
optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

Examples

library(scda)
library(dplyr)

adsl <- synthetic_cdisc_data("latest")$adsl
adex <- synthetic_cdisc_data("latest")$adex

set.seed(1, kind = "Mersenne-Twister")
labels <- var_labels(adex)
adex <- adex %>%
 distinct(USUBJID, .keep_all = TRUE) %>%
  mutate(PARAMCD = "TDURD",
         PARAM = "Overall duration (days)",
         AVAL = sample(x = seq(1, 200), size = n(), replace = TRUE),
         AVALU = "Days") %>%
         bind_rows(adex)
 var_labels(adex) <- labels

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", adsl, code = 'ADSL <- synthetic_cdisc_data("latest")$adsl'),
    cdisc_dataset("ADEX", adex,
      code = 'set.seed(1, kind = "Mersenne-Twister")
      ADEX <- synthetic_cdisc_data("latest")$adex
      labels <- var_labels(ADEX)
      ADEX <- ADEX %>%
       distinct(USUBJID, .keep_all = TRUE) %>%
       mutate(PARAMCD = "TDURD",
              PARAM = "Overall duration (days)",
              AVAL = sample(x = seq(1, 200), size = n(), replace = TRUE),
              AVALU = "Days") %>%
              bind_rows(ADEX)
      var_labels(ADEX) <- labels'  #nolint
    ),
    check = TRUE
  ),
  modules = root_modules(
    tm_t_exposure(
      label = "Duration of Exposure Table",
      dataname = "ADEX",
      paramcd = choices_selected(
        choices = value_choices(adex, "PARAMCD", "PARAM"),
        selected = "TDURD"
      ),
      col_by_var = choices_selected(
        choices = variable_choices(adex, subset = c("SEX", "ARM")),
        selected = "SEX"
      ),
      row_by_var = choices_selected(
        choices = variable_choices(adex, subset = c("RACE", "REGION1", "STRATA1", "SEX")),
        selected = "RACE"
      ),
      parcat = choices_selected(
        choices = value_choices(adex, "PARCAT2"),
        selected = "Drug A"
      ),
      add_total = FALSE
    )
  ),
 filter = list(
   ADSL = list(SAFFL = "Y")
 )
)
#> 2021-11-29 15:34:49.1540 pid:962 teal.modules.clinical fun:tm_t_exposure [INFO] Initializing tm_t_exposure

if (FALSE) {
shinyApp(app$ui, app$server)
}