Teal Module: Patient Profile Laboratory Teal Module

This teal module produces a patient profile laboratory table using ADaM datasets.

tm_t_pp_laboratory(
  label,
  dataname = "ADLB",
  parentname = "ADSL",
  patient_col = "USUBJID",
  timepoints = NULL,
  aval = NULL,
  avalu = NULL,
  param = NULL,
  paramcd = NULL,
  anrind = NULL,
  pre_output = NULL,
  post_output = NULL
)

Arguments

label

(character)
menu item label of the module in the teal app.

dataname

(character)
analysis data used in teal module.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

patient_col

(character)
patient ID column to be used.

timepoints

(teal::choices_selected() or teal::data_extract_spec())
Time variable to be represented in the laboratory table.

aval

(teal::choices_selected() or teal::data_extract_spec())
AVAL column of the ADLB dataset.

avalu

(teal::choices_selected() or teal::data_extract_spec())
AVALU column of the ADLB dataset.

param

(teal::choices_selected() or teal::data_extract_spec())
PARAM column of the ADLB dataset.

paramcd

(teal::choices_selected() or teal::data_extract_spec())
PARAMCD column of the ADLB dataset.

anrind

(teal::choices_selected() or teal::data_extract_spec())
ANRIND column of the ADLB dataset with 3 possible levels "HIGH", "LOW" and "NORMAL".

pre_output

optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output

optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

Examples

library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
ADLB <- synthetic_cdisc_data("latest")$adlb

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- synthetic_cdisc_data('latest')$adsl"),
    cdisc_dataset("ADLB", ADLB, code = 'ADLB <- synthetic_cdisc_data("latest")$adlb'),
    check = TRUE
  ),
  modules = root_modules(
    tm_t_pp_laboratory(
      label = "Vitals",
      dataname = "ADLB",
      patient_col = "USUBJID",
      paramcd = choices_selected(
        choices = variable_choices(ADLB, "PARAMCD"),
        selected = "PARAMCD"
      ),
      param = choices_selected(
        choices = variable_choices(ADLB, "PARAM"),
        selected = "PARAM"
      ),
      timepoints = choices_selected(
        choices = variable_choices(ADLB, "ADY"),
        selected = "ADY"
      ),
      anrind = choices_selected(
        choices = variable_choices(ADLB, "ANRIND"),
        selected = "ANRIND"
      ),
      aval = choices_selected(
        choices = variable_choices(ADLB, "AVAL"),
        selected = "AVAL"
      ),
      avalu = choices_selected(
        choices = variable_choices(ADLB, "AVALU"),
        selected = "AVALU"
      )
    )
  )
)
#> 2022-01-04 07:43:51.2547 pid:2129 token: teal.modules.clinical fun:tm_t_pp_laboratory [INFO] Initializing tm_t_pp_laboratory
if (FALSE) {
shinyApp(app$ui, app$server)
}