Teal Module: Patient Medical History Teal Module

This teal module produces a patient medical history report using ADaM datasets.

tm_t_pp_medical_history(
  label,
  dataname = "ADMH",
  parentname = "ADSL",
  patient_col = "USUBJID",
  mhterm = NULL,
  mhbodsys = NULL,
  mhdistat = NULL,
  pre_output = NULL,
  post_output = NULL
)

Arguments

label

(character)
menu item label of the module in the teal app.

dataname

(character)
analysis data used in teal module.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

patient_col

(character)
patient ID column to be used.

mhterm

(teal::choices_selected() or teal::data_extract_spec())
MHTERM column of the ADMH dataset.

mhbodsys

(teal::choices_selected() or teal::data_extract_spec())
MHBODSYS column of the ADMH dataset.

mhdistat

(teal::choices_selected() or teal::data_extract_spec())
MHDISTAT column of the ADMH dataset.

pre_output

optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output

optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

Examples

library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
ADMH <- synthetic_cdisc_data("latest")$admh
ADMH[["MHDISTAT"]] <- "ONGOING"
rtables::var_labels(ADMH[c("MHDISTAT")]) <- c("Status of Disease")

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = 'ADSL <- synthetic_cdisc_data("latest")$adsl'),
    cdisc_dataset("ADMH", ADMH, code = "ADMH <- synthetic_cdisc_data('latest')$admh
                   ADMH[['MHDISTAT']] <- 'ONGOING'
                   rtables::var_labels(ADMH[c('MHDISTAT')]) <- c('Status of Disease')"),
    check = TRUE
  ),
  modules = root_modules(
    tm_t_pp_medical_history(
      label = "Medical history",
      dataname = "ADMH",
      parentname = "ADSL",
      patient_col = "USUBJID",
      mhterm = choices_selected(
        choices = variable_choices(ADMH, c("MHTERM")),
        selected = "MHTERM"
      ),
      mhbodsys = choices_selected(
        choices = variable_choices(ADMH, "MHBODSYS"),
        selected = "MHBODSYS"
      ),
      mhdistat = choices_selected(
        choices = variable_choices(ADMH, "MHDISTAT"),
        selected = "MHDISTAT"
      )
    )
  )
)
#> [INFO] 2022-02-10 16:28:20.4481 pid:2150 token:[] teal.modules.clinical Initializing tm_t_pp_medical_history
if (FALSE) {
shinyApp(app$ui, app$server)
}