Response Table Teal Module

This module produces a response summary table that matches the STREAM template rspt01

tm_t_rsp(label, dataname, arm_var, arm_ref_comp = NULL, paramcd,
  strata_var, pre_output = NULL, post_output = NULL,
  code_data_processing = NULL)

Arguments

label
menue item label of the module in the teal app

dataname
analysis data used in teal module, needs to be available in the list passed to the data argument of init. Note that the data is expected to be in vertical form with the PARAMCD variable filtering to one observation per patient.

arm_var
choices_selected object with all available choices and preselected option for variable names that can be used as arm_var

arm_ref_comp
optional, if specified it must be a named list with each element corresponding to an arm variable in ASL and the element must be another list with the elements named ref and comp that the defined the default reference and comparison arms when the arm variable is changed.

paramcd
choices_selected object with all available choices and preselected option for variable names that can be used as PARAMCD variable

strata_var
choices_selected object with all available choices and preselected option for variable names that can be used for stratification

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

code_data_processing
string with data preprocessing before the teal app is initialized

Value

an module object

Details

Additional standard UI inputs include responders, ref_arm, comp_arm and combine_arm (default FALSE)

Default values of the inputs var_arm, ref_arm and comp_arm are set to NULL, and updated accordingly based on seletion of paramcd and var_arm

This display order of response categories in partitioned statistics section inherits the factor level order of the source data. Use factor and its levels argument to manipulate the source data in order to include/exclude or re-categorize response categories and arrange the display order. If response categories are "Missing" or "Not Evaluable (NE)" or "Missing or unevaluable", 95% confidence interval will not be calculated.

Reference arms automatically combined if multiple arms selected as reference group.

Examples

library(random.cdisc.data)

ASL <- radsl(seed = 1)
ARS <- subset(radrs(ASL, seed = 1), AVISIT == "Follow Up")

attr(ASL, "source") <- "random.cdisc.data::radsl(seed = 1)"
attr(ARS, "source") <- 'subset(random.cdisc.data::radrs(ASL, seed = 1), AVISIT == "Follow Up")'

x <- teal::init(
  data = list(ASL = ASL, ARS = ARS),
  modules = root_modules(
    tm_t_rsp(
       label = "Response Table",
       dataname = 'ARS',
       arm_var = choices_selected(c("ARM", "ARMCD"), "ARM"),
       paramcd = choices_selected(unique(ARS$PARAMCD), "BESRSPI"),
       strata_var = choices_selected(c("SEX", "BMRKR2"), "SEX")
    )
  )
)

#> |apply filter for ASL
#> |apply filter for ARS
#> |** UI id for modules is teal_modules.root
#> |UI id for module is teal_modules_root.Response_Table

# NOT RUN {
shinyApp(x$ui, x$server)

# }