Teal Module: Summarize Variables by Row Groups Module

Teal Module: Summarize Variables by Row Groups Module

tm_t_summary_by(
  label,
  dataname,
  parentname = ifelse(inherits(arm_var, "data_extract_spec"),
    teal.transform::datanames_input(arm_var), "ADSL"),
  arm_var,
  by_vars,
  summarize_vars,
  id_var = teal.transform::choices_selected(teal.transform::variable_choices(dataname,
    subset = "USUBJID"), selected = "USUBJID", fixed = TRUE),
  paramcd = NULL,
  add_total = TRUE,
  parallel_vars = FALSE,
  row_groups = FALSE,
  useNA = c("ifany", "no"),
  na_level = "<Missing>",
  numeric_stats = c("n", "mean_sd", "median", "range"),
  denominator = teal.transform::choices_selected(c("n", "N", "omit"), "omit", fixed =
    TRUE),
  drop_arm_levels = TRUE,
  drop_zero_levels = TRUE,
  pre_output = NULL,
  post_output = NULL,
  basic_table_args = teal.widgets::basic_table_args()
)

Arguments

label

(character)
menu item label of the module in the teal app.

dataname

(character)
analysis data used in teal module.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var

(choices_selected or data_extract_spec)
object with all available choices and preselected option for variable names that can be used as arm_var. It defines the grouping variable(s) in the results table. If there are two elements selected for arm_var, second variable will be nested under the first variable.

by_vars

(choices_selected or data_extract_spec)
object with all available choices and preselected option for variable names used to split the summary by rows.

summarize_vars

(choices_selected or data_extract_spec)
names of the variables that should be summarized.

id_var

(choices_selected or data_extract_spec)
object specifying the variable name for subject id.

paramcd

(choices_selected or data_extract_spec)
variable value designating the studied parameter.

add_total

(logical)
whether to include column with total number of patients.

parallel_vars

(logical) used to display summarize_vars as parallel columns (FALSE on default). Can be used only if all chosen analysis variables are numeric.

row_groups

(logical) used to display summarize_vars as row groups (FALSE on default).

useNA

(character)
whether missing data (NA) should be displayed as a level.

na_level

(string)
used to replace all NA or empty values in character or factor variables in the data.

numeric_stats

(character)
selected statistics for numeric summarize variables to be displayed. Possible values are n, mean_sd, mean_ci, median, median_ci, range. By default, n, mean_sd, median, range are selected.

denominator

(character)
chooses how percentages are calculated. With option N, the reference population from the column total is used as the denominator. With option n, the number of non-missing records in this row and column intersection is used as the denominator. If omit is chosen, then the percentage is omitted.

drop_arm_levels

(logical)
drop the unused arm_var levels. When TRUE, arm_var levels are set to those used in the dataname dataset. When FALSE, arm_var levels are set to those used in the parentname dataset. If dataname dataset and parentname dataset are the same (i.e. ADSL), then drop_arm_levels will always be TRUE regardless of the user choice when tm_t_summary_by is called.

drop_zero_levels

(logical) used to remove rows with zero counts from the result table.

pre_output

optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output

optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

basic_table_args

optional, (basic_table_args)
object created by teal.widgets::basic_table_args() with settings for the module table. The argument is merged with option teal.basic_table_args and with default module arguments (hard coded in the module body).

For more details, see the vignette: vignette("custom-basic-table-arguments", package = "teal.widgets").

Examples

# Preparation of the test case.
library(scda)
adsl <- synthetic_cdisc_data("latest")$adsl
adlb <- synthetic_cdisc_data("latest")$adlb

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", adsl, code = 'ADSL <- synthetic_cdisc_data("latest")$adsl'),
    cdisc_dataset("ADLB", adlb, code = 'ADLB <- synthetic_cdisc_data("latest")$adlb'),
    check = TRUE
  ),
  modules = modules(
    tm_t_summary_by(
      label = "Summary by Row Groups Table",
      dataname = "ADLB",
      arm_var = choices_selected(
        choices = variable_choices(adsl, c("ARM", "ARMCD")),
        selected = "ARM"
      ),
      add_total = TRUE,
      by_vars = choices_selected(
        choices = variable_choices(adlb, c("PARAM", "AVISIT")),
        selected = c("AVISIT")
      ),
      summarize_vars = choices_selected(
        choices = variable_choices(adlb, c("AVAL", "CHG")),
        selected = c("AVAL")
      ),
      useNA = "ifany",
      paramcd = choices_selected(
        choices = value_choices(adlb, "PARAMCD", "PARAM"),
        selected = "ALT"
      )
    )
  )
)
#> [INFO] 2022-04-11 00:24:29.0916 pid:2753 token:[] teal.modules.clinical Initializing tm_t_summary_by
if (FALSE) {
shinyApp(app$ui, app$server)
}