ggplot2 call to generate simple barchart

ggplot2 call to generate simple barchart

Usage

make_barchart_simple_call(
  y_name,
  x_name = NULL,
  fill_name = NULL,
  x_facet_name = NULL,
  y_facet_name = NULL,
  label_bars = TRUE,
  barlayout = c("side_by_side", "stacked"),
  flip_axis = FALSE,
  rotate_bar_labels = FALSE,
  rotate_x_label = FALSE,
  rotate_y_label = FALSE,
  expand_y_range = 0,
  facet_scales = "free_x",
  ggplot2_args = teal.widgets::ggplot2_args()
)

Arguments

fill_name

(NULL, character(1))
name of the variable distinguishing the color of the fill.

x_facet_name

(NULL, character(1))
name of the variable to facet the plot horizontally.

y_facet_name

(NULL, character(1))
name of the variable to facet the plot vertically.

label_bars

(NULL, logical(1))
whether to label bars.

barlayout

(NULL, character(1))
type of the layout - "side_by_side" or "stacked" (default).

flip_axis

(NULL, character(1))
whether to flip axis.

rotate_bar_labels

(NULL, logical(1))
whether to rotate bar label by 45 degrees.

rotate_x_label

(NULL, logical(1))
whether to rotate x-axis label by 45 degrees.

rotate_y_label

(NULL, logical(1))
whether to rotate y-axis label by 45 degrees.

expand_y_range

(NULL, numeric(1))
fraction of y-axis range to expand further up.

facet_scales

(fixed, free_x, free_y or free) value passed to scales argument of ggplot2::facet_grid

ggplot2_args

optional, (ggplot2_args)
object created by teal.widgets::ggplot2_args() with settings for the module plot. The argument is merged with option teal.ggplot2_args and with default module arguments (hard coded in the module body).

For more details, see the help vignette: vignette("custom-ggplot2-arguments", package = "teal.widgets").

yname

(NULL, character(1))
name of the yaxis variable.

xname

(NULL, character(1))
name of the xaxis variable.

Value

call

Examples

teal.modules.clinical:::make_barchart_simple_call(y_name = "y", x_name = "x")
#> plot <- ggplot2::ggplot(counts) + ggplot2::aes_string(x = "x") + 
#>     ggplot2::geom_col(ggplot2::aes_string(y = "y"), position = ggplot2::position_dodge(0.9), 
#>         fill = "#b6cae9") + ggplot2::geom_text(ggplot2::aes_string(y = "y", 
#>     label = "y"), stat = "identity", angle = 0, position = ggplot2::position_dodge(0.9), 
#>     hjust = 0.5, vjust = -1) + ggplot2::labs(x = column_annotation_label(counts, 
#>     "x"))