Standard Module Arguments

The documentation to this function lists all the arguments in teal modules that are used repeatedly to express an analysis.

Arguments

arm_ref_comp

optional, (list)
If specified it must be a named list with each element corresponding to an arm variable in ADSL and the element must be another list (possibly with delayed teal.transform::variable_choices() or delayed teal.transform::value_choices() with the elements named ref and comp that the defined the default reference and comparison arms when the arm variable is changed.

arm_var

(teal.transform::choices_selected())
object with all available choices and preselected option for variable names that can be used as arm_var. It defines the grouping variable(s) in the results table. If there are two elements selected for arm_var, second variable will be nested under the first variable.

atirel

(teal.transform::choices_selected())
object with all available choices and preselected option for the ATIREL variable from dataname.

aval_var

(teal.transform::choices_selected())
object with all available choices and pre-selected option for the analysis variable.

avalu_var

(teal.transform::choices_selected())
object with all available choices and preselected option for the analysis unit variable.

avisit

(teal.transform::choices_selected())
value of analysis visit AVISIT of interest.

baseline_var

(teal.transform::choices_selected())
object with all available choices and preselected option for variable values that can be used as baseline_var.

by_vars

(teal.transform::choices_selected())
object with all available choices and preselected option for variable names used to split the summary by rows.

cmdecod

(teal.transform::choices_selected())
object with all available choices and preselected option for the CMDECOD variable from dataname.

cmindc

(teal.transform::choices_selected())
object with all available choices and preselected option for the CMINDC variable from dataname.

cmstdy

(teal.transform::choices_selected())
object with all available choices and preselected option for the CMSTDY variable from dataname.

cnsr_var

(teal.transform::choices_selected())
object with all available choices and preselected option for the censoring variable.

conf_level

(teal.transform::choices_selected())
object with all available choices and pre-selected option for the confidence level, each within range of (0, 1).

cov_var

(teal.transform::choices_selected())
object with all available choices and preselected option for the covariates variables.

dataname

(character)
analysis data used in teal module.

default_responses

(list or character)
defines the default codes for the response variable in the module per value of paramcd. A passed vector is transmitted for all paramcd values. A passed list must be named and contain arrays, each name corresponding to a single value of paramcd. Each array may contain default response values or named arrays rsp of default selected response values and levels of default level choices.

fixed_symbol_size

(logical)
When (TRUE), the same symbol size is used for plotting each estimate. Otherwise, the symbol size will be proportional to the sample size in each each subgroup.

font_size

(numeric)
numeric vector of length 3 of current, minimum and maximum font size values.

hlt

(teal.transform::choices_selected())
name of the variable with high level term for events.

id_var

(teal.transform::choices_selected())
object specifying the variable name for subject id.

interact_var

(character)
name of the variable that should have interactions with arm. If the interaction is not needed, the default option is NULL.

interact_y

(character)
a selected item from the interact_var column which will be used to select the specific ANCOVA results when interact_var is discrete. If the interaction is not needed, the default option is FALSE.

label

(character)
menu item label of the module in the teal app.

llt

(teal.transform::choices_selected())
name of the variable with low level term for events.

paramcd

(teal.transform::choices_selected())
object with all available choices and preselected option for the parameter code variable from dataname.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

patient_col

(character)
name of patient ID variable.

plot_height

optional, (numeric)
a vector of length three with c(value, min, max). Specifies the height of the main plot and renders a slider on the plot to interactively adjust the plot height.

plot_width

optional, (numeric)
a vector of length three with c(value, min, max). Specifies the width of the main plot and renders a slider on the plot to interactively adjust the plot width.

post_output

optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

pre_output

optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

strata_var

(teal.transform::choices_selected())
names of the variables for stratified analysis.

summarize_vars

(teal.transform::choices_selected())
names of the variables that should be summarized.

subgroup_var

(teal.transform::choices_selected())
object with all available choices and preselected option for variable names that can be used as the default subgroups.

time_points

(teal.transform::choices_selected())
object with all available choices and preselected option for time points that can be used in tern::surv_timepoint().

time_unit_var

(teal.transform::choices_selected())
object with all available choices and pre-selected option for the time unit variable.

treatment_flag

(teal.transform::choices_selected())
value indicating on treatment records in treatment_flag_var.

treatment_flag_var

(teal.transform::choices_selected())
on treatment flag variable.

useNA

(character)
whether missing data (NA) should be displayed as a level.

visit_var

(teal.transform::choices_selected())
object with all available choices and preselected option for variable names that can be used as visit variable. Must be a factor in dataname.

worst_flag_indicator

(teal.transform::choices_selected())
value indicating worst grade.

worst_flag_var

(teal.transform::choices_selected())
object with all available choices and preselected option for variable names that can be used as worst flag variable.

Value

a teal_module object.

Details

Although this function just returns NULL it has two uses, for the teal module users it provides a documentation of arguments that are commonly and consistently used in the framework. For the developer it adds a single reference point to import the roxygen argument description with: @inheritParams module_arguments

Parameters with identical descriptions & input types to those in the Standard Template Arguments section are excluded to reduce duplication as each module function inherits parameters from its corresponding template function.

See also