Template: Adverse Events Tab

Creates an adverse events template call.

template_adverse_events(
  dataname = "ANL",
  aeterm = "AETERM",
  tox_grade = "AETOXGR",
  causality = "AEREL",
  outcome = "AEOUT",
  action = "AEACN",
  time = "ASTDY",
  decod = NULL,
  patient_id,
  font_size = 12L
)

Arguments

dataname
(character)
analysis data used in teal module.

aeterm
(character)
name of the reported term for the adverse event variable.

tox_grade
(character)
name of the standard toxicity grade variable.

causality
(character)
name of the causality variable.

outcome
(character)
name of outcome of adverse event variable.

action
(character)
name of action taken with study treatment variable.

time
(character)
name of study day of start of adverse event variable.

decod
(character)
name of dictionary derived term variable.

patient_id
(character)
patient ID.

font_size
(numeric)
numeric vector of length 3 for current, min and max font size values.