Template: Adverse Events Tab

Creates an adverse events template call.

template_adverse_events(
  dataname = "ae_merge",
  ae_term = "AETERM",
  ae_tox_grade = "AETOXGR",
  ae_causality = "AEREL",
  ae_outcome = "AEOUT",
  ae_action = "AEACN",
  ae_time = "ASTDY",
  ae_decod = NULL,
  font_size = 12L
)

Arguments

dataname
(character)
analysis data used in teal module.

ae_term
(character)
name of the reported term for the adverse event variable.

ae_tox_grade
(character)
name of the standard toxicity grade variable.

ae_causality
(character)
name of the causality variable.

ae_outcome
(character)
name of outcome of adverse event variable.

ae_action
(character)
name of action taken with study treatment variable.

ae_time
(character)
name of study day of start of adverse event variable.

ae_decod
(choices selected or data_extract_input)
AEDECOD column of the ADAE dataset.

font_size
(numeric)
numeric vector of length 3 for current, min and max font size values.