Standard Template Arguments

The documentation to this function lists all the arguments in teal module templates that are used repeatedly to express an analysis.

Arguments

arm_var

(character)
variable names that can be used as arm_var.

id_var

(character)
the variable name for subject id.

sum_vars

(character)
names of the variables that should be summarized.

compare_arm

(logical)
triggers the comparison between study arms.

combine_comp_arms

(logical)
triggers the combination of comparison arms.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

dataname

(character)
analysis data used in teal module.

anl_name

(character)
analysis data used in teal module.

paramcd

(character)
variable value designating the studied parameter.

ref_arm

(character)
the level of reference arm in case of arm comparison.

comp_arm

(character)
the level of comparison arm in case of arm comparison.

na.rm

(logical)
whether NA values should be removed prior to analysis.

useNA

(character)
whether missing data (NA) should be displayed as a level.

na_level

(string)
used to replace all NA or empty values in character or factor variables in the data.

denominator

(character)
chooses how percentages are calculated. With option N, the reference population from the column total is used as the denominator. With option n, the number of non-missing records in this row and column intersection is used as the denominator. If omit is chosen, then the percentage is omitted.

var_labels

optional, (named character)
variable labels for relabeling the analysis variables.

cov_var

(character)
names of the covariates variables.

aval_var

(character)
name of the analysis variable.

cnsr_var

(character)
name of the censoring variable.

conf_level

(numeric)
value for the confidence level within the range of (0, 1).

add_total

(logical)
whether to include column with total number of patients.

hlt

(character)
name of the variable with high level term for events.

llt

(character)
name of the variable with low level term for events.

visit_var

(character)
variable names that can be used as visit variable. Must be a factor in dataname.

by_vars

(character)
variable names used to split the summary by rows.

time_points

(character)
time points that can be used in tern::surv_timepoint().

time_unit_var

(character)
name of the variable representing time units.

drop_arm_levels

(logical)
drop the unused arm_var levels. When TRUE, arm_var levels are set to those used in the dataname dataset. When FALSE, arm_var levels are set to those used in the parantname dataset.

subgroup_var

(character)
with variable names that can be used as subgroups.

strata_var

(character)
names of the variables for stratified analysis.

title

(character)
title of the output.

prune_freq

(number)
threshold to use for trimming table using event incidence rate in any column.

prune_diff

(number)
threshold to use for trimming table using as criteria difference in rates between any two columns.

treatment_flag_var

(character)
name of the on treatment flag variable.

treatment_flag

(character)
name of the value indicating on treatment records in treatment_flag_var.

basic_table_args

optional, (basic_table_args)
object created by teal.widgets::basic_table_args() with settings for the module table. The argument is merged with option teal.basic_table_args and with default module arguments (hard coded in the module body).

For more details, see the vignette: vignette("custom-basic-table-arguments", package = "teal.widgets").

ggplot2_args

optional, (ggplot2_args)
object created by teal.widgets::ggplot2_args() with settings for the module plot. The argument is merged with option teal.ggplot2_args and with default module arguments (hard coded in the module body).

For more details, see the vignette: vignette("custom-ggplot2-arguments", package = "teal.widgets").

Details

Although this function just returns NULL it has two uses, for the teal module users it provides a documentation of arguments that are commonly and consistently used in the framework. For the developer it adds a single reference point to import the roxygen argument description with: @inheritParams template_arguments