Template: Adverse Events grouped by Grade with threshold

Template: Adverse Events grouped by Grade with threshold

Usage

template_events_col_by_grade(
  dataname,
  parentname,
  arm_var,
  grading_groups = list(`Any Grade (%)` = c("1", "2", "3", "4", "5"), `Grade 1-2 (%)` =
    c("1", "2"), `Grade 3-4 (%)` = c("3", "4"), `Grade 5 (%)` = "5"),
  add_total = TRUE,
  id = "USUBJID",
  hlt,
  llt,
  label_hlt = NULL,
  label_llt = NULL,
  grade = "AETOXGR",
  label_grade = NULL,
  prune_freq = 0.1,
  prune_diff = 0,
  drop_arm_levels = TRUE,
  basic_table_args = teal.widgets::basic_table_args()
)

Arguments

dataname

(character)
analysis data used in teal module.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var

(character)
variable names that can be used as arm_var.

grading_groups

(character)
list of grading groups.

add_total

(logical)
whether to include column with total number of patients.

id

(character)
unique identifier of patients in datasets, default to "USUBJID".

hlt

(character)
name of the variable with high level term for events.

llt

(character)
name of the variable with low level term for events.

label_hlt

(string)
label of the hlt variable from dataname. The label will be extracted from the module.

label_llt

(string)
label of the llt variable from dataname. The label will be extracted from the module.

grade

(character)
grade term which grading_groups is based on, default to "AETOXGR".

label_grade

(string)
label of the grade variable from dataname. The label will be extracted from the module.

prune_freq

(number)
threshold to use for trimming table using event incidence rate in any column.

prune_diff

(number)
threshold to use for trimming table using as criteria difference in rates between any two columns.

drop_arm_levels

(logical)
drop the unused arm_var levels. When TRUE, arm_var levels are set to those used in the dataname dataset. When FALSE, arm_var levels are set to those used in the parantname dataset.

basic_table_args

optional, (basic_table_args)
object created by teal.widgets::basic_table_args() with settings for the module table. The argument is merged with option teal.basic_table_args and with default module arguments (hard coded in the module body).

For more details, see the vignette: vignette("custom-basic-table-arguments", package = "teal.widgets").

See also