Template: Response Forest Plot

Creates a valid expression for response forest plot.

template_forest_rsp(
  dataname = "ANL",
  parentname = "ADSL_FILTERED",
  arm_var,
  ref_arm = NULL,
  comp_arm = NULL,
  aval_var = "AVALC",
  responders = c("CR", "PR"),
  subgroup_var,
  strata_var = NULL,
  conf_level = 0.95,
  col_symbol_size = NULL
)

Arguments

dataname
(character)
analysis data used in teal module.

parentname
(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var
(character)
variable names that can be used as arm_var.

ref_arm
(character)
the level of reference arm in case of arm comparison.

comp_arm
(character)
the level of comparison arm in case of arm comparison.

aval_var
(character)
name of the analysis variable.

responders
(character)
values of aval_var that are considered to be responders.

subgroup_var
(character)
with variable names that can be used as subgroups.

strata_var
(character)
names of the variables for stratified analysis.

conf_level
(numeric)
value for the confidence level within the range of (0, 1).

col_symbol_size
(integer)
column index to be used to determine relative size for estimator plot symbol. Typically, the symbol size is proportional to the sample size used to calculate the estimator. If NULL, the same symbol size is used for all subgroups.

See also