Template: Individual Patient Plots

Template: Individual Patient Plots

Usage

template_g_ipp(
  dataname = "ANL",
  paramcd,
  arm_var,
  arm_levels,
  avalu_first,
  paramcd_first,
  aval_var = "AVAL",
  avalu_var = "AVALU",
  id_var = "USUBJID",
  visit_var = "AVISIT",
  base_var = "BASE",
  add_baseline_hline = FALSE,
  separate_by_obs = FALSE,
  ggplot2_args = teal.widgets::ggplot2_args(),
  suppress_legend = FALSE,
  add_avalu = TRUE
)

Arguments

dataname

(character)
analysis data used in teal module.

paramcd

(character)
variable value designating the studied parameter.

arm_var

(character)
variable names that can be used as arm_var.

arm_levels

(character)
vector of all arm variable levels.

avalu_first

(string)
avalu value.

paramcd_first

(string)
paramcd value.

aval_var

(character)
name of the analysis variable.

avalu_var

(string)
variable name designating the unit of the analysis variable.

id_var

(character)
the variable name for subject id.

visit_var

(string)
variable name designating the visit timepoint variable.

base_var

(string)
variable name designating the baseline values of analysis variable.

add_baseline_hline

(flag)
adds horizontal line at baseline y-value on plot

separate_by_obs

(flag)
creates multi panel plots when TRUE

ggplot2_args

optional, (ggplot2_args)
object created by teal.widgets::ggplot2_args() with settings for the module plot. For this module, this argument will only accept ggplot2_args object with labs list of following child elements: title, subtitle, x, y. No other elements would be taken into account. The argument is merged with option teal.ggplot2_args and with default module arguments (hard coded in the module body).

For more details, see the vignette: vignette("custom-ggplot2-arguments", package = "teal.widgets").

suppress_legend

(flag)
allow user to suppress legend

add_avalu

(flag)
allow user to not display value unit in the plot.