Simple barchart plot that shows counts per category.

Categories can be defined up to four levels deep and are defined through x, fill, x_facet and y_facet. If any of them is NULL, it is ignored.

tm_g_barchart_simple(
  x = NULL,
  fill = NULL,
  x_facet = NULL,
  y_facet = NULL,
  label = "Count Barchart",
  plot_options = NULL,
  plot_height = c(600L, 200L, 2000L),
  plot_width = NULL,
  pre_output = NULL,
  post_output = NULL
)

Arguments

x
(data_extract_spec)
variable on the x-axis.

fill
(data_extract_spec)
grouping variable assigned to the bar colors.

x_facet
(data_extract_spec)
faceting groups on the row dimension.

y_facet
(data_extract_spec)
faceting groups on the col dimension.

label
(character)
menu item label of the module in the teal app

plot_options
(list)
list of plot options.

plot_height
optional, (numeric)
a vector of length three with c(value, min, max). Specifies the height of the main plot and renders a slider on the plot to interactively adjust the plot height.

plot_width
optional, (numeric)
a vector of length three with c(value, min, max). Specifies the width of the main plot and renders a slider on the plot to interactively adjust the plot width.

pre_output
optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output
optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

Examples

library(dplyr)
library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl
ADAE <- synthetic_cdisc_data("latest")$adae

adae_labels <- var_labels(ADAE)
ADAE <- ADAE %>% dplyr::filter(!((AETOXGR == 1) & (AESEV == "MILD") & (ARM == "A: Drug X")))

# reinstate labels
ADAE <- do.call(rtables::var_relabel, append(list(x = ADAE), as.list(adae_labels)))

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = "ADSL <- synthetic_cdisc_data('latest')$adsl"),
    cdisc_dataset("ADAE", ADAE,
      code = "ADAE <- synthetic_cdisc_data('latest')$adae
              adae_labels <- var_labels(ADAE)
              ADAE <- ADAE %>%
                dplyr::filter(!((AETOXGR == 1) & (AESEV == 'MILD') & (ARM == 'A: Drug X')))
              ADAE <- do.call(rtables::var_relabel,
                append(list(x = ADAE), as.list(adae_labels)))"),
    check = TRUE
  ),
  modules = root_modules(
    tm_g_barchart_simple(
      label = "ADAE Analysis",
      x = data_extract_spec(
        dataname = "ADSL",
        select = select_spec(
          choices = variable_choices(ADSL,
                                     c("ARM", "ACTARM", "SEX",
                                       "RACE", "ITTFL", "SAFFL", "STRATA2")
          ),
          selected = "ACTARM",
          multiple = FALSE
        )
      ),
      fill = list(
        data_extract_spec(
          dataname = "ADSL",
          select = select_spec(
            choices = variable_choices(ADSL,
                                       c("ARM", "ACTARM", "SEX",
                                         "RACE", "ITTFL", "SAFFL", "STRATA2")
            ),
            selected = "SEX",
            multiple = FALSE
          )
        ),
        data_extract_spec(
          dataname = "ADAE",
          select = select_spec(
            choices = variable_choices(ADAE, c("AETOXGR", "AESEV", "AESER")),
            selected = NULL,
            multiple = FALSE
          )
        )
      ),
      x_facet = list(
        data_extract_spec(
          dataname = "ADAE",
          select = select_spec(
            choices = variable_choices(ADAE, c("AETOXGR", "AESEV", "AESER")),
            selected = "AETOXGR",
            multiple = FALSE
          )
        ),
        data_extract_spec(
          dataname = "ADSL",
          select = select_spec(
            choices = variable_choices(ADSL,
                                       c("ARM", "ACTARM", "SEX",
                                         "RACE", "ITTFL", "SAFFL", "STRATA2")
            ),
            selected = NULL,
            multiple = FALSE
          )
        )
      ),
      y_facet = list(
        data_extract_spec(
          dataname = "ADAE",
          select = select_spec(
            choices = variable_choices(ADAE, c("AETOXGR", "AESEV", "AESER")),
            selected = "AESEV",
            multiple = FALSE
          )
        ),
        data_extract_spec(
          dataname = "ADSL",
          select = select_spec(
            choices = variable_choices(ADSL,
                                       c("ARM", "ACTARM", "SEX",
                                         "RACE", "ITTFL", "SAFFL", "STRATA2")
            ),
            selected = NULL,
            multiple = FALSE
          )
        )
      )
    )
  )
)
if (FALSE) {
shinyApp(app$ui, app$server)
}