Forest Survival Plot teal Module

This is teal module produces a grid style Forest plot for time-to-event data with ADaM structure

tm_g_forest_tte(label, dataname, arm_var, subgroup_var, paramcd,
  plot_height = c(600, 200, 2000), cex = 1.3,
  pre_output = helpText("graph needs to be of a certain width to be displayed"),
  post_output = NULL)

Arguments

label
menu item label of the module in the teal app

dataname
(character) analysis data used in teal module, needs to be available in the list passed to the data argument of init. Note that the data is expected to be in vertical form with the PARAMCD variable filtering to one observation per patient.

arm_var
choices_selected object with all available choices and preselected option for variable names that can be used as arm_var

subgroup_var
choices_selected object with all available choices and preselected option for variable names that can be used as the default subgroups

paramcd
choices_selected object with all available choices and preselected option for variable names that can be used as PARAMCD variable

plot_height
vector with three elements defining selected, min and max plot height

cex
multiplier applied to overall font size

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

Examples

library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADTTE <- radtte(ADSL, cached = TRUE)

ADSL$RACE <- droplevels(ADSL$RACE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADTTE", ADTTE),
    code = 'ADSL <- ADSL <- radsl(cached = TRUE)
            ADTTE <- radtte(ADSL, cached = TRUE)
            ADSL$RACE <- droplevels(ADSL$RACE)',
    check = FALSE
    ),
  modules = root_modules(
    tm_g_forest_tte(
       label = "Forest Survival",
       dataname = "ADTTE",
       arm_var = choices_selected(c("ARM", "ARMCD"), "ARM"),
       subgroup_var = choices_selected(names(ADSL), c("RACE", "SEX")),
       paramcd = choices_selected(c("OS", "PFS"), "OS"),
       plot_height = c(600, 200, 2000)
    )
  )
)

if (FALSE) {
  shinyApp(app$ui, app$server)
}