Teal Module: Individual Patient Plot

This teal module produces grid style Individual patient plot(s) that show trends in parameter values over time for each patient using data with ADaM structure.

tm_g_ipp(
  label,
  dataname,
  parentname = ifelse(is(arm_var, "data_extract_spec"), datanames_input(arm_var),
    "ADSL"),
  arm_var,
  paramcd,
  id_var = choices_selected(variable_choices(dataname, "USUBJID"), "USUBJID", fixed =
    TRUE),
  visit_var = choices_selected(variable_choices(dataname, "AVISIT"), "AVISIT", fixed =
    TRUE),
  aval_var = choices_selected(variable_choices(dataname, "AVAL"), "AVAL", fixed = TRUE),
  avalu_var = choices_selected(variable_choices(dataname, "AVALU"), "AVALU", fixed =
    TRUE),
  base_var = choices_selected(variable_choices(dataname, "BASE"), "BASE", fixed = TRUE),
  add_baseline_hline = FALSE,
  separate_by_obs = FALSE,
  plot_height = c(1200L, 400L, 5000L),
  plot_width = NULL,
  pre_output = NULL,
  post_output = NULL
)

Arguments

label
(character)
menu item label of the module in the teal app

dataname
(character)
analysis data used in teal module.

parentname
(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable values that can be used as arm_var.

paramcd
(teal::choices_selected() or teal::data_extract_spec())
variable value designating the studied parameter.

id_var
(teal::choices_selected() or teal::data_extract_spec())
object specifying the variable name for subject id.

visit_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable names that can be used as visit variable. Must be a factor in dataname.

aval_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for the analysis variable.

avalu_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable values that can be used as avalu_var.

base_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable values that can be used as base_var.

add_baseline_hline
(flag)
adds horizontal line at baseline y-value on plot

separate_by_obs
(flag)
creates multi panel plots when TRUE

plot_height
optional, (numeric)
a vector of length three with c(value, min, max). Specifies the height of the main plot and renders a slider on the plot to interactively adjust the plot height.

plot_width
optional, (numeric)
a vector of length three with c(value, min, max). Specifies the width of the main plot and renders a slider on the plot to interactively adjust the plot width.

pre_output
optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output
optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

Examples

library(dplyr)
library(random.cdisc.data)

adsl <- radsl(cached = TRUE) %>% slice(1:20)
adlb <- radlb(adsl, seed = 3)

adsl <- df_explicit_na(adsl)
adlb <- df_explicit_na(adlb) %>%
 dplyr::filter(AVISIT != "SCREENING")

app <- init(
  data = cdisc_data(
    cdisc_dataset(
    "ADSL",
    adsl,
    code = "ADSL <- radsl(cached = TRUE) %>% slice(1:20)
            ADSL <- df_explicit_na(ADSL)"
            ),
    cdisc_dataset(
    "ADLB",
    adlb,
    code = "ADLB <- radlb(ADSL, seed = 3)
            ADLB <- df_explicit_na(ADLB) %>%
              dplyr::filter(AVISIT != 'SCREENING')"
             )
  ),
  modules = root_modules(
    tm_g_ipp(
      label = "Individual Patient Plot",
      dataname = "ADLB",
      arm_var = choices_selected(
        value_choices(adlb, "ARMCD"),
        "ARM A"
      ),
      paramcd = choices_selected(
        value_choices(adlb, "PARAMCD"),
        "ALT"
      ),
      aval_var = choices_selected(
        variable_choices(adlb, c("AVAL", "CHG")),
        "AVAL"
      ),
      avalu_var = choices_selected(
        variable_choices(adlb, c("AVALU")),
        "AVALU",
        fixed = TRUE
      ),
      id_var = choices_selected(
        variable_choices(adlb, c("USUBJID")),
        "USUBJID",
        fixed = TRUE
      ),
      visit_var = choices_selected(
        variable_choices(adlb, c("AVISIT")),
        "AVISIT"
      ),
       base_var = choices_selected(
        variable_choices(adlb, c("BASE")),
        "BASE",
        fixed = TRUE
      ),
      add_baseline_hline = FALSE,
      separate_by_obs = FALSE
      )
    )
  )

if (FALSE) {
shinyApp(ui = app$ui, server = app$server)
}