teal module for Kaplan-Meier Plot from grid

This is teal module produces a grid style KM plot for data with ADaM structure

tm_g_km(label, dataname, arm_var, arm_ref_comp = NULL, paramcd,
  facet_var, strata_var, plot_height = c(1200, 400, 5000),
  tbl_fontsize = 8,
  pre_output = helpText("x-axes for different factes may not have the same scale"),
  post_output = NULL, code_data_processing = NULL)

Arguments

label
menue item label of the module in the teal app

dataname
analysis data used in teal module, needs to be available in the list passed to the data argument of init. Note that the data is expected to be in vertical form with the PARAMCD variable filtering to one observation per patient.

arm_var
choices_selected object with all available choices and preselected option for variable names that can be used as arm_var

arm_ref_comp
optional, if specified it must be a named list with each element corresponding to an arm variable in ASL and the element must be another list with the elements named ref and comp that the defined the default reference and comparison arms when the arm variable is changed.

paramcd
choices_selected object with all available choices and preselected option for variable names that can be used as PARAMCD variable

facet_var
choices_selected object with all available choices and preselected option for variable names that can be used for facet plotting

strata_var
choices_selected object with all available choices and preselected option for variable names that can be used for stratification

plot_height
vector with three elements defining selected, min and max plot height

tbl_fontsize
fontsize for text annotation

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

code_data_processing
string with data preprocessing before the teal app is initialized

Examples

library(random.cdisc.data)

ASL <- radsl(seed = 1)
ATE <- radtte(ASL, seed = 1)

attr(ASL, "source") <- "random.cdisc.data::radsl(seed = 1)"
attr(ATE, "source") <- "random.cdisc.data::radtte(ASL, seed = 1)"

arm_ref_comp = list(
   ARM = list(
      ref = "A: Drug X",
      comp = c("B: Placebo", "C: Combination")
   ),
   ARMCD = list(
      ref = "ARM B",
      comp = "ARM A"
   )
)

x <- teal::init(
  data = list(ASL = ASL, ATE = ATE),
  modules = root_modules(
    tm_g_km(
       label = "KM PLOT",
       dataname = 'ATE',
       arm_var = choices_selected(c("ARM", "ARMCD"), "ARM"),
       arm_ref_comp = arm_ref_comp,
       paramcd = choices_selected(c("OS", "PFS"), "OS"),
       facet_var = choices_selected(c("SEX", "BMRKR2"), "BMRKR2"),
       strata_var = choices_selected(c("SEX", "BMRKR2"), "SEX"),
       tbl_fontsize = 12
    )
  )
)

#> |apply filter for ASL
#> |apply filter for ATE
#> |** UI id for modules is teal_modules.root
#> |UI id for module is teal_modules_root.KM_PLOT

# NOT RUN {
shinyApp(ui = x$ui, server = x$server)

# }