Teal Module: Line Plot

This teal module produces a grid style Line Plot for data with ADaM structure.

tm_g_lineplot(
  label,
  dataname,
  parentname = ifelse(is(strata, "data_extract_spec"), datanames_input(strata), "ADSL"),
  strata = choices_selected(variable_choices(parentname, c("ARM", "ARMCD",
    "ACTARMCD")), "ARM"),
  x = choices_selected(variable_choices(dataname, "AVISIT"), "AVISIT", fixed = TRUE),
  y = choices_selected(variable_choices(dataname, c("AVAL", "BASE", "CHG", "PCHG")),
    "AVAL"),
  y_unit = choices_selected(variable_choices(dataname, "AVALU"), "AVALU", fixed = TRUE),
  paramcd = choices_selected(variable_choices(dataname, "PARAMCD"), "PARAMCD", fixed =
    TRUE),
  param = choices_selected(value_choices(dataname, "PARAMCD", "PARAM"), "ALT"),
  conf_level = choices_selected(c(0.95, 0.9, 0.8), 0.95, keep_order = TRUE),
  interval = "mean_ci",
  mid = "mean",
  whiskers = c("mean_ci_lwr", "mean_ci_upr"),
  table = NULL,
  mid_type = "pl",
  mid_point_size = c(2, 1, 5),
  table_font_size = c(4, 2, 6),
  plot_height = c(1000L, 200L, 4000L),
  plot_width = NULL,
  pre_output = NULL,
  post_output = NULL
)

Arguments

label
(character)
menu item label of the module in the teal app

dataname
(character)
analysis data used in teal module.

parentname
(character)
parent analysis data used in teal module, usually this refers to ADSL.

strata
(character scalar or NA) strata variable name.

x
(character scalar) x variable name.

y
(character scalar) y variable name.

y_unit
(character scalar or NA) y_unit variable name.

paramcd
(character scalar or NA) paramcd variable name.

param
(character)
parameter chosen to filter the data by.

conf_level
(numeric)
value for the confidence level within the range of (0, 1).

interval
(character or NULL)
names of the statistics that will be plotted as intervals. All the statistics indicated in interval variable must be present in the object returned by sfun, and be of a double or numeric type vector of length two.

mid
(character or NULL)
names of the statistics that will be plotted as midpoints. All the statistics indicated in mid variable must be present in the object returned by sfun, and be of a double or numeric type vector of length one.

whiskers
(character)
names of the interval whiskers that will be plotted. Must match the names attribute of the interval element in the list returned by sfun. It is possible to specify one whisker only, lower or upper.

table
(character or NULL)
names of the statistics that will be displayed in the table below the plot. All the statistics indicated in table variable must be present in the object returned by sfun.

mid_type
(character scalar)
controls the type of the mid plot, it can be point (p), line (l), or point and line (pl).

mid_point_size
(integer or double)
controls the font size of the point for mid plot.

table_font_size
(integer or double)
controls the font size of values in the table.

plot_height
optional, (numeric)
a vector of length three with c(value, min, max). Specifies the height of the main plot and renders a slider on the plot to interactively adjust the plot height.

plot_width
optional, (numeric)
a vector of length three with c(value, min, max). Specifies the width of the main plot and renders a slider on the plot to interactively adjust the plot width.

pre_output
optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output
optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

Examples

library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl
ADLB <- synthetic_cdisc_data("latest")$adlb

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL, code = 'ADSL <- synthetic_cdisc_data("latest")$adsl'),
    cdisc_dataset("ADLB", ADLB, code = 'ADLB <- synthetic_cdisc_data("latest")$adlb'),
    check = TRUE
  ),
  modules = root_modules(
    tm_g_lineplot(
      label = "Line Plot",
      dataname = "ADLB",
      strata = choices_selected(
        variable_choices(ADSL, c("ARM", "ARMCD", "ACTARMCD")),
        "ARM"
      ),
      y = choices_selected(
        variable_choices(ADLB, c("AVAL", "BASE", "CHG", "PCHG")),
        "AVAL"
      ),
      param = choices_selected(
        value_choices(ADLB, "PARAMCD", "PARAM"),
        "ALT"
      )
    )
  )
)

if (FALSE) {
shinyApp(ui = app$ui, server = app$server)
}