Teal Module: Patient Profile Adverse Events Teal Module

This teal module produces a patient profile adverse events table and plot using ADaM datasets.

Usage

tm_g_pp_adverse_events(
  label,
  dataname = "ADAE",
  parentname = "ADSL",
  patient_col = "USUBJID",
  aeterm = NULL,
  tox_grade = NULL,
  causality = NULL,
  outcome = NULL,
  action = NULL,
  time = NULL,
  decod = NULL,
  font_size = c(12L, 12L, 25L),
  plot_height = c(700L, 200L, 2000L),
  plot_width = NULL,
  pre_output = NULL,
  post_output = NULL,
  ggplot2_args = teal.widgets::ggplot2_args()
)

Arguments

label

(character)
menu item label of the module in the teal app.

dataname

(character)
analysis data used in teal module.

parentname

(character)
parent analysis data used in teal module, usually this refers to ADSL.

patient_col

(character)
patient ID column to be used.

aeterm

(teal.transform::choices_selected() or teal.transform::data_extract_spec())
AETERM column of the ADAE dataset.

tox_grade

(teal.transform::choices_selected() or teal.transform::data_extract_spec())
AETOXGR column of the ADAE dataset.

causality

(teal.transform::choices_selected() or teal.transform::data_extract_spec())
AEREL column of the ADAE dataset.

outcome

(teal.transform::choices_selected() or teal.transform::data_extract_spec())
AEOUT column of the ADAE dataset.

action

(teal.transform::choices_selected() or teal.transform::data_extract_spec())
AEACN column of the ADAE dataset.

time

(teal.transform::choices_selected() or teal.transform::data_extract_spec())
ASTDY column of the ADAE dataset.

decod

(teal.transform::choices_selected() or teal.transform::data_extract_spec())
AEDECOD column of the ADAE dataset.

font_size

(numeric)
numeric vector of length 3 for current, min and max font size values.

plot_height

optional, (numeric)
a vector of length three with c(value, min, max). Specifies the height of the main plot and renders a slider on the plot to interactively adjust the plot height.

plot_width

optional, (numeric)
a vector of length three with c(value, min, max). Specifies the width of the main plot and renders a slider on the plot to interactively adjust the plot width.

pre_output

optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output

optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

ggplot2_args

optional, (ggplot2_args)
object created by teal.widgets::ggplot2_args() with settings for the module plot. The argument is merged with option teal.ggplot2_args and with default module arguments (hard coded in the module body).

For more details, see the help vignette: vignette("custom-ggplot2-arguments", package = "teal.widgets").

Examples

library(scda)
library(nestcolor)

synthetic_cdisc_data_latest <- synthetic_cdisc_data("latest")
ADSL <- synthetic_cdisc_data_latest$adsl
ADAE <- synthetic_cdisc_data_latest$adae

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL,
      code = "synthetic_cdisc_data_latest <- synthetic_cdisc_data('latest')
        ADSL <- synthetic_cdisc_data_latest$adsl"
    ),
    cdisc_dataset("ADAE", ADAE,
      code = "synthetic_cdisc_data_latest <- synthetic_cdisc_data('latest')
        ADAE <- synthetic_cdisc_data_latest$adae"
    )
  ),
  modules = modules(
    tm_g_pp_adverse_events(
      label = "Adverse events",
      dataname = "ADAE",
      parentname = "ADSL",
      patient_col = "USUBJID",
      plot_height = c(600L, 200L, 2000L),
      aeterm = choices_selected(
        choices = variable_choices(ADAE, "AETERM"),
        selected = "AETERM"
      ),
      tox_grade = choices_selected(
        choices = variable_choices(ADAE, "AETOXGR"),
        selected = "AETOXGR"
      ),
      causality = choices_selected(
        choices = variable_choices(ADAE, "AEREL"),
        selected = "AEREL"
      ),
      outcome = choices_selected(
        choices = variable_choices(ADAE, "AEOUT"),
        selected = "AEOUT"
      ),
      action = choices_selected(
        choices = variable_choices(ADAE, "AEACN"),
        selected = "AEACN"
      ),
      time = choices_selected(
        choices = variable_choices(ADAE, "ASTDY"),
        selected = "ASTDY"
      ),
      decod = NULL
    )
  )
)
#> [INFO] 2022-10-14 09:10:06.3959 pid:3139 token:[] teal.modules.clinical Initializing tm_g_pp_adverse_events
if (FALSE) {
shinyApp(app$ui, app$server)
}