This module produces a summary table that includes abnormality counts based on t_abnormality. For correct results, please use the filter panel to select only post-baseline records.

This module produces a summary table that includes abnormality counts based on t_abnormality. For correct results, please use the filter panel to select only post-baseline records.

tm_t_abnormality(
  label,
  dataname,
  arm_var,
  id_var,
  by_vars,
  grade,
  abnormal,
  baseline,
  exclude_base_abn = FALSE,
  pre_output = NULL,
  post_output = NULL
)

Arguments

label
menu item label of the module in the teal app

dataname
(character) analysis data used in teal module, needs to be available in the list passed to the data argument of init.

arm_var
choices_selected object with all available choices and preselected option for variable names that can be used as arm

id_var
choices_selected object specifying the variable name for subject id.

by_vars
choices_selected object with all available choices and preselected option for variable names used to split the summary by rows.

grade
choices_selected object with all available choices and preselected option for variable names that can be used to specify the abnormality grade. Variable must be factor.

abnormal
(vector) to specify the default preselected values indicating abnormality grade.

baseline
choices_selected object with all available choices and preselected option for variable names that can be used to specify the baseline abnormality.

exclude_base_abn
(boolean) to specify the default option in UI whether to exclude subjects with baseline abnormality in denominator. Default is FALSE.

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example a title.

post_output
optional, shiny.tag with text placed after the output to put the output into context. For example the helpText elements are useful.

Value

a module object

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ADLB0 <- radlb(cached = TRUE)
# create BNRIND
ADLB_base <- ADLB0 %>%
  dplyr::filter(ABLFL == "Y") %>%
  dplyr::select(USUBJID, PARAM, ANRIND) %>%
  dplyr::rename(BNRIND = ANRIND)

ADLB <- merge(ADLB0, ADLB_base, by = c("USUBJID", "PARAM")) %>%
  var_relabel(
    PARAM = "Parameters",
    BNRIND = "Baseline Reference Range Indicator",
    ANRIND = "Analysis Reference Range Indicator",
    AVISIT = "Analysis Visit"
  )

app <- init(
data = cdisc_data(
  cdisc_dataset("ADSL", ADSL),
  cdisc_dataset("ADLB", ADLB),
  code = 'ADSL <- radsl(cached = TRUE)
         ADLB0 <- radlb(cached = TRUE)
         ADLB_base <- ADLB0 %>%
           dplyr::filter(ABLFL == "Y") %>%
           dplyr::select(USUBJID, PARAM, ANRIND) %>%
           dplyr::rename(BNRIND = ANRIND)
         ADLB <- merge(ADLB0, ADLB_base, by = c("USUBJID", "PARAM")) %>%
           var_relabel(
             PARAM = "Parameters",
             ANRIND = "Analysis Reference Range Indicator",
             AVISIT = "Analysis Visit")',
  check = FALSE),
 modules = root_modules(
   tm_t_abnormality(
     label = "Abnormality Table",
     dataname = 'ADLB',
     arm_var = choices_selected(
       choices = variable_choices(ADSL, subset = c("ARM", "ARMCD")),
       selected = "ARM"
     ),
     id_var = choices_selected(
       choices = variable_choices(ADSL, subset = c("USUBJID", "SUBJID")),
       selected = "USUBJID",
       fixed = TRUE
     ),
     by_vars = choices_selected(
       choices = variable_choices(ADLB, subset = c("LBCAT", "PARAM", "AVISIT")),
       selected = c("PARAM", "AVISIT"),
       keep_order = TRUE
     ),
     grade = choices_selected(
       choices = variable_choices(ADLB, subset = "ANRIND"),
       selected = "ANRIND",
       fixed = TRUE
     ),
     abnormal = c("LOW", "HIGH"),
     baseline = choices_selected(
       choices = variable_choices(ADLB, subset = "BNRIND"),
       selected = "BNRIND",
       fixed = TRUE
     ),
     exclude_base_abn = FALSE
   )
 )
)

# important to filter out baseline records for table numbers to be correct
app$datasets$hold_filtering()
app$datasets$set_filter_state("ADLB", "ABLFL", "")
app$datasets$set_filter_state("ADLB", "ABLFL2", "")
app$datasets$continue_filtering()

if (FALSE) {
shinyApp(app$ui, app$server)
}