Teal Module: ANCOVA Teal Module

Teal Module: ANCOVA Teal Module

tm_t_ancova(
  label,
  dataname,
  parentname = ifelse(is(arm_var, "data_extract_spec"), datanames_input(arm_var),
    "ADSL"),
  arm_var,
  arm_ref_comp = NULL,
  aval_var,
  cov_var,
  avisit,
  paramcd,
  conf_level = choices_selected(c(0.95, 0.9, 0.8), 0.95, keep_order = TRUE),
  pre_output = NULL,
  post_output = NULL
)

Arguments

label
(character)
menu item label of the module in the teal app

dataname
(character)
analysis data used in teal module.

parentname
(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable names that can be used as arm_var.

arm_ref_comp
optional, (list)
If specified it must be a named list with each element corresponding to an arm variable in ADSL and the element must be another list (possibly with delayed teal::variable_choices() or delayed teal::value_choices() with the elements named ref and comp that the defined the default reference and comparison arms when the arm variable is changed.

aval_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for the analysis variable.

cov_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for the covariates variables.

avisit
(teal::choices_selected() or teal::data_extract_spec())
value of analysis visit AVISIT of interest.

paramcd
(teal::choices_selected() or teal::data_extract_spec())
variable value designating the studied parameter.

conf_level
(teal::choices_selected())
object with all available choices and preselected option for the confidence level, each within range of (0, 1).

pre_output
optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output
optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

Details

This module produces an analysis of variance summary table that is similar to STREAM template aovt01 when multiple endpoints are selected. When a single endpoint is selected, both unadjusted and adjusted comparison would be provided. This modules expects that the analysis data has the following variables:

AVISIT	variable used to filter for analysis visits.

PARAMCD	variable used to filter for endpoints, after filtering for paramcd and avisit, one observation per patient is expected for the analysis to be meaningful.

Examples

# Preparation of the test case.
library(dplyr)
library(random.cdisc.data)
adsl <- radsl(cached = TRUE)
adqs <- radqs(cached = TRUE)

arm_ref_comp = list(
  ARM = list(
    ref = "B: Placebo",
    comp = c("A: Drug X", "C: Combination")
  ),
  ACTARMCD = list(
    ref = "ARM B",
    comp = c("ARM A", "ARM C")
  )
)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", adsl),
    cdisc_dataset("ADQS", adqs),
    code =
      "ADSL <- radsl(cached = TRUE)
       ADQS <- radqs(cached = TRUE)"
  ),
  modules = root_modules(
    tm_t_ancova(
      label = "ANCOVA table",
      dataname = "ADQS",
      avisit = choices_selected(
        choices = value_choices(adqs, "AVISIT"),
        selected = "WEEK 1 DAY 8"
      ),
      arm_var = choices_selected(
        choices = variable_choices(adsl, c("ARM", "ACTARMCD")),
        selected = "ARMCD"
      ),
      arm_ref_comp = arm_ref_comp,
      aval_var = choices_selected(
        choices = variable_choices(adqs, c("CHG", "AVAL")),
        selected = "CHG"
      ),
      cov_var = choices_selected(
        choices = variable_choices(adqs, c("BASE", "STRATA1", "SEX")),
        selected = "STRATA1"
      ),
      paramcd = choices_selected(
        choices = value_choices(adqs, "PARAMCD", "PARAM"),
        selected = "FKSI-FWB"
      )
    )
  )
)

if (FALSE) {
shinyApp(app$ui, app$server)
}