Cox Multiple Regression Model

Teal module to program the simple effects of a multi-variable Cox regression models such as: Surv(time, event) ~ Arm + Cov1 + ... . This provides the COXT02 standard output.

tm_t_coxreg(
  label,
  dataname,
  arm_var,
  arm_ref_comp = NULL,
  paramcd,
  cov_var,
  strata_var,
  conf_level = choices_selected(c(0.8, 0.85, 0.9, 0.95, 0.99, 0.995), 0.95, keep_order =
    TRUE),
  pre_output = NULL,
  post_output = NULL
)

Arguments

label
Menu item label of the module in the teal app.

dataname
character analysis data used in teal module, needs to be available in the list passed to the data argument of init. Note that the data is expected to be in vertical form with the PARAMCD variable filtering to one observation per patient.

arm_var
choices_selected object with all available choices and preselected option for variable names that can be used as arm_var

arm_ref_comp
(choices_selected) optional, if specified it must be a named list with each element corresponding to an arm variable in ADSL and the element must be another list with the elements named ref and comp that the defined the default reference and comparison arms when the arm variable is changed.

paramcd
choices_selected object with all available choices and preselected option for variable names that can be used as PARAMCD variable.

cov_var
choices_selected object with all available choices and preselected option for variable names that can be used as covariates in the model.

strata_var
choices_selected object with all available choices and preselected option for variable names that can be used for stratification

conf_level
choices_selected object with all available choices and preselected option for variable names that can be used for confidence level for computation of the confidence intervals.

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example a title.

post_output
optional, shiny.tag with text placed after the output to put the output into context. For example the helpText elements are useful.

Details

This modules expects that the analysis data has the following variables

AVAL	time to event

CNSR	boolean or 0,1 is element in AVAL censored

PARAMCD	variable used to filter for endpoint (e.g. OS), after filtering for paramcd one observation per patient is expected

The arm variables, stratification and covariate variables are taken from the ADSL data.

Note

The likelihood ratio test is not supported for model including strata, Wald test will be substituted.

Examples

library(random.cdisc.data)

ADSL  <- radsl(cached = TRUE)
ADTTE <- radtte(cached = TRUE)
arm_ref_comp = list(
  ACTARMCD = list(
    ref = "ARM B",
    comp = c("ARM A", "ARM C")
  ),
  ARM = list(
    ref = "B: Placebo",
    comp = c("A: Drug X", "C: Combination")
  )
)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADTTE", ADTTE),
    code = "ADSL  <- radsl(cached = TRUE);  ADTTE <- radtte(cached = TRUE)"
  ),
  modules = root_modules(
    tm_t_coxreg(
      label = "Cox Reg.",
      dataname = "ADTTE",
      arm_var = choices_selected(c("ARM", "ARMCD", "ACTARMCD"), "ARM"),
      arm_ref_comp = arm_ref_comp,
      paramcd = choices_selected(value_choices(ADTTE, "PARAMCD", "PARAM"), "OS"),
      strata_var = choices_selected(c("COUNTRY", "STRATA1", "STRATA2"), "STRATA1"),
      cov_var    = choices_selected(
        c("AGE", "BMRKR1", "BMRKR2", "REGION1"), "AGE"
      )
    )
  )
)
if (FALSE) {
shinyApp(app$ui, app$server)
}