Teal Module: Grade Summary Table

Teal Module: Grade Summary Table

tm_t_shift_by_grade(
  label,
  dataname,
  parentname = ifelse(is(arm_var, "data_extract_spec"), datanames_input(arm_var),
    "ADSL"),
  arm_var,
  visit_var = choices_selected(variable_choices(dataname, subset = "AVISIT"), selected
    = "AVISIT", fixed = TRUE),
  paramcd,
  worst_flag_var = choices_selected(variable_choices(dataname, subset = c("WGRLOVFL",
    "WGRLOFL", "WGRHIVFL", "WGRHIFL")), selected = "WGRLOVFL"),
  worst_flag_indicator = choices_selected(value_choices(dataname, "WGRLOVFL"), selected
    = "Y", fixed = TRUE),
  anl_toxgrade_var = choices_selected(variable_choices(dataname, subset = c("ATOXGR")),
    selected = c("ATOXGR"), fixed = TRUE),
  base_toxgrade_var = choices_selected(variable_choices(dataname, subset =
    c("BTOXGR")), selected = c("BTOXGR"), fixed = TRUE),
  id_var = choices_selected(variable_choices(dataname, subset = "USUBJID"), selected =
    "USUBJID", fixed = TRUE),
  add_total = FALSE,
  drop_arm_levels = TRUE,
  pre_output = NULL,
  post_output = NULL,
  na_level = "<Missing>",
  code_missing_baseline = FALSE
)

Arguments

label
(character)
menu item label of the module in the teal app

dataname
(character)
analysis data used in teal module.

parentname
(character)
parent analysis data used in teal module, usually this refers to ADSL.

arm_var
(teal::choices_selected() or teal::data_extract_spec())
object with all available choices and preselected option for variable names that can be used as arm_var. It defines the grouping variable(s) in the results table. If there are two elements selected for arm_var, second variable will be nested under the first variable.

visit_var
(teal::choices_selected() or teal::data_extract_spec)
object with all available choices and preselected option for variable names that can be used as visit.

paramcd
(teal::choices_selected() or teal::data_extract_spec())
variable value designating the studied parameter.

worst_flag_var
(teal::choices_selected() or teal::data_extract_spec)
object with all available choices and preselected option for variable names that can be used as worst flag variable.

worst_flag_indicator
(teal::choices_selected() or teal::data_extract_spec)
value indicating worst grade.

anl_toxgrade_var
(teal::choices_selected() or teal::data_extract_spec)
variable for analysis toxicity grade.

base_toxgrade_var
(teal::choices_selected() or teal::data_extract_spec)
variable for baseline toxicity grade.

id_var
(teal::choices_selected() or teal::data_extract_spec())
object specifying the variable name for subject id.

add_total
(logical)
whether to include column with total number of patients.

drop_arm_levels
(logical)
drop the unused arm_var levels. When TRUE, arm_var levels are set to those used in the dataname dataset. When FALSE, arm_var levels are set to those used in the parantname dataset.

pre_output
optional, (shiny.tag)
with text placed before the output to put the output into context. For example a title.

post_output
optional, (shiny.tag)
with text placed after the output to put the output into context. For example the shiny::helpText() elements are useful.

na_level
(string)
used to replace all NA or empty values in character or factor variables in the data.

code_missing_baseline
(character)
whether missing baseline should be considered as grade 0.

Examples

library(scda)
library(dplyr)

adsl <- synthetic_cdisc_data("latest")$adsl
adlb <- synthetic_cdisc_data("latest")$adlb

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", adsl, code = 'ADSL <- synthetic_cdisc_data("latest")$adsl'),
    cdisc_dataset("ADLB", adlb,
                  code = 'ADLB <- synthetic_cdisc_data("latest")$adlb'),
    check = TRUE
  ),
  modules = root_modules(
    tm_t_shift_by_grade(
      label = "Grade Laboratory Abnormality Table",
      dataname = "ADLB",
      arm_var = choices_selected(
        choices = variable_choices(adsl, subset = c("ARM", "ARMCD")),
        selected = "ARM"
      ),
      paramcd = choices_selected(
        choices = value_choices(adlb, "PARAMCD", "PARAM"),
        selected = "ALT"
      ),
      worst_flag_var = choices_selected(
        choices = variable_choices(adlb, subset = c("WGRLOVFL", "WGRLOFL", "WGRHIVFL", "WGRHIFL")),
        selected = c("WGRLOVFL")
      ),
      worst_flag_indicator = choices_selected(
        value_choices(adlb, "WGRLOVFL"), selected = "Y", fixed = TRUE
      ),
      anl_toxgrade_var = choices_selected(
        choices = variable_choices(adlb, subset = c("ATOXGR")),
        selected = c("ATOXGR"),
        fixed = TRUE
      ),
      base_toxgrade_var = choices_selected(
        choices = variable_choices(adlb, subset = c("BTOXGR")),
        selected = c("BTOXGR"),
        fixed = TRUE
      ),
      add_total = FALSE
   )
 ),
  filter = list(
    ADSL = list(SAFFL = "Y")
  )
)
#> 2021-12-10 13:46:06.0763 pid:868 teal.modules.clinical fun:tm_t_shift_by_grade [INFO] Initializing tm_t_shift_by_grade

 if (FALSE) {
shinyApp(app$ui, app$server)
}