Summarize variables with by group module

This module produces a summary table that includes row-by variables based on t_summary_by.

tm_t_summary_by(
  label,
  dataname,
  arm_var,
  by_vars,
  summarize_vars,
  parallel_vars = FALSE,
  useNA = c("ifany", "no"),
  denominator = c("n", "N", "omit"),
  pre_output = NULL,
  post_output = NULL
)

Arguments

label
menu item label of the module in the teal app

dataname
(character) analysis data used in teal module, needs to be available in the list passed to the data argument of init.

arm_var
choices_selected object with all available choices and preselected option for variable names that can be used as arm_var

by_vars
choices_selected object with all available choices and preselected option for variable names used to split the summary by rows.

summarize_vars
\link[teal]{choices_selected} object with all available choices and preselected option for variable names that can be used for summary

parallel_vars
(logical) used to display summarize_vars as parallel columns (FALSE on default). Can be used only if all chosen analysis variables are numeric.

useNA
default option in UI whether to display NA values, see t_summary

denominator
default option in UI for denominator, for calculating percentages. Only applies to categorical variables. See "denominator" in t_summary.factor for details.

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example a title.

post_output
optional, shiny.tag with text placed after the output to put the output into context. For example the helpText elements are useful.

Value

a module object

Examples

library(random.cdisc.data)
library(dplyr)

ADSL <- radsl(cached = TRUE)
ADLB <- radlb(cached = TRUE) %>%
mutate(
   CHGCAT1 = case_when(
     CHG > 0 ~ "INCREASE",
     CHG < 0 ~ "DECREASE",
     CHG == 0 ~ "NO CHANGE"
   )
 ) %>%
 var_relabel(CHGCAT1 = "Change from Baseline Category 1")

# CHG in ADLB contains NA values so useNA can be tested
stopifnot(all(vapply(ADLB[c("CHG")], function(x) any(is.na(x)), logical(1))))

app <- init(
 data = cdisc_data(
   cdisc_dataset("ADSL", ADSL),
   cdisc_dataset("ADLB", ADLB),
   code = 'ADSL <- radsl(cached = TRUE)
           ADLB <- radlb(cached = TRUE) %>%
            mutate(
               CHGCAT1 = case_when(
                 CHG > 0 ~ "INCREASE",
                 CHG < 0 ~ "DECREASE",
                 CHG == 0 ~ "NO CHANGE"
               )
             ) %>%
             var_relabel(CHGCAT1 = "Change from Baseline Category 1")',
   check = FALSE
 ),
 modules = root_modules(
   tm_t_summary_by(
     label = "By-Variable Table",
     dataname = "ADLB",
     arm_var = choices_selected(
       choices = variable_choices(ADSL, c("ARM", "ARMCD")),
       selected = "ARM"
     ),
     by_vars = choices_selected(
       choices = variable_choices(ADLB, c("PARAM", "AVISIT")),
       selected = c("PARAM", "AVISIT")
     ),
     summarize_vars = choices_selected(
       choices = variable_choices(ADLB, c("AVAL", "CHG", "CHGCAT1")),
       selected = c("AVAL")
     ),
     denominator = "N",
     useNA = "ifany"
   )
 )
)

if (FALSE) {
shinyApp(app$ui, app$server)
}